@@ -8,7 +8,7 @@ readme = "README.md"
 keywords = ["interpreter", "language-programming", "tacit", "symbolic"]
 categories = ["science", "parser-implementations", "internationalization", "compression"]
 license = "GPL-3.0"
-edition="2018"
+edition="2021"
 
 [badges]
 maintenance = {status = "actively-developed"}
@@ -25,7 +25,7 @@ dashmap = "3.5.0"
 lazy_static = "1.2.0"
 maplit = "1.0.2"
 rayon = "1.3.0"
-snafu = "0.5.0"
+thiserror = "1"
 log = { version = "0.4", features = ["max_level_debug", "release_max_level_warn"] }
 slotmap = "1.0.6"
 
@@ -35,3 +35,4 @@ simple_logger = "1.15.0"
 
 [dev-dependencies]
 proptest = "0.9.1"
+pretty_assertions = "1.4.0"
