@@ -2,11 +2,11 @@ use crate::{
     ast::SyntaxTree,
     concepts::{ConceptTrait, ConcreteConceptType, Hand},
     context_delta::{
-        Composition, ConceptDelta, ContextDelta, DirectConceptDelta,
-        NewDirectConceptDelta, ValueChange,
+        Composition, ConceptDelta, DirectConceptDelta, NestedDelta,
+        NewDirectConceptDelta, SharedDelta, ValueChange,
     },
-    context_snap_shot::Uncommitted,
     errors::{ZiaError, ZiaResult},
+    mixed_concept::MixedConcept,
 };
 use std::{
     convert::TryFrom,
@@ -18,18 +18,12 @@ pub trait Reader<SDCD>
 where
     SDCD: Clone
         + AsRef<DirectConceptDelta<Self::ConceptId>>
-        + From<DirectConceptDelta<Self::ConceptId>>,
+        + From<DirectConceptDelta<Self::ConceptId>>
+        + Debug,
     Self::CommittedConceptId: TryFrom<Self::ConceptId>,
     Self::ConceptId: From<Self::CommittedConceptId>,
 {
-    type ConceptId: Copy
-        + Eq
-        + Hash
-        + Display
-        + Debug
-        + Send
-        + Sync
-        + From<Uncommitted>;
+    type ConceptId: MixedConcept + Send + Sync;
     type CommittedConceptId: Copy + Eq + Hash + Display + Debug + Send + Sync;
     type MixedConcept<'a>: ConceptTrait<Id = Self::ConceptId>
         + Clone
@@ -43,21 +37,23 @@ where
         &self,
         concept_id: Self::ConceptId,
     ) -> Option<Self::MixedConcept<'_>>;
-    fn read_concept<'a>(
+    fn read_concept<
+        'a,
+        D: SharedDelta<NestedDelta = NestedDelta<Self::ConceptId, SDCD, D>>,
+    >(
         &'a self,
-        delta: &ContextDelta<Self::ConceptId, SDCD>,
+        delta: &NestedDelta<Self::ConceptId, SDCD, D>,
         id: Self::ConceptId,
     ) -> Self::MixedConcept<'a> {
         delta
-            .concept()
-            .get(&id)
+            .get_concept(&id)
             .and_then(|cds| {
                 let mut concept = self.get_concept(id);
                 for cd in cds {
                     match cd {
                         ConceptDelta::Direct(dcd) => match dcd.as_ref() {
                             DirectConceptDelta::New(delta) => {
-                                debug_assert!(concept.is_none());
+                                debug_assert!(concept.is_none()); // Assert that concept hasn't already been committed
                                 concept = Some((&NewDirectConceptDelta{
                                     delta: delta.clone(),
                                     new_concept_id: id
@@ -140,12 +136,15 @@ where
             })
             .unwrap_or_else(|| {
                 self.get_concept(id)
-                    .unwrap_or_else(|| panic!("No concept with id = {}", id))
+                    .unwrap_or_else(|| panic!("No concept with id = {id}"))
             })
     }
-    fn concepts_from_composition<'a>(
+    fn concepts_from_composition<
+        'a,
+        D: SharedDelta<NestedDelta = NestedDelta<Self::ConceptId, SDCD, D>>,
+    >(
         &'a self,
-        delta: &ContextDelta<Self::ConceptId, SDCD>,
+        delta: &NestedDelta<Self::ConceptId, SDCD, D>,
         comp: Composition<Self::ConceptId>,
     ) -> [Self::MixedConcept<'a>; 2] {
         [
@@ -153,14 +152,19 @@ where
             self.read_concept(delta, comp.right_id),
         ]
     }
-    fn get_label(
+    fn get_label<
+        D: SharedDelta<NestedDelta = NestedDelta<Self::ConceptId, SDCD, D>>,
+    >(
         &self,
-        delta: &ContextDelta<Self::ConceptId, SDCD>,
+        delta: &NestedDelta<Self::ConceptId, SDCD, D>,
         concept_id: Self::ConceptId,
     ) -> Option<String>;
-    fn ast_from_symbol<Syntax>(
+    fn ast_from_symbol<
+        Syntax,
+        D: SharedDelta<NestedDelta = NestedDelta<Self::ConceptId, SDCD, D>>,
+    >(
         &self,
-        delta: &ContextDelta<Self::ConceptId, SDCD>,
+        delta: &NestedDelta<Self::ConceptId, SDCD, D>,
         s: &str,
     ) -> Syntax
     where
@@ -174,9 +178,12 @@ where
             },
         )
     }
-    fn bind_concept_to_syntax<Syntax>(
+    fn bind_concept_to_syntax<
+        Syntax,
+        D: SharedDelta<NestedDelta = NestedDelta<Self::ConceptId, SDCD, D>>,
+    >(
         &self,
-        delta: &ContextDelta<Self::ConceptId, SDCD>,
+        delta: &NestedDelta<Self::ConceptId, SDCD, D>,
         syntax: Syntax,
         concept: Self::ConceptId,
     ) -> Syntax
@@ -208,14 +215,18 @@ where
             SyntaxTree::new_constant_concept(concept.id())
         }
     }
-    fn concept_from_label(
+    fn concept_from_label<
+        D: SharedDelta<NestedDelta = NestedDelta<Self::ConceptId, SDCD, D>>,
+    >(
         &self,
-        delta: &ContextDelta<Self::ConceptId, SDCD>,
+        delta: &NestedDelta<Self::ConceptId, SDCD, D>,
         s: &str,
     ) -> Option<Self::ConceptId>;
-    fn get_reduction_of_composition(
+    fn get_reduction_of_composition<
+        D: SharedDelta<NestedDelta = NestedDelta<Self::ConceptId, SDCD, D>>,
+    >(
         &self,
-        delta: &ContextDelta<Self::ConceptId, SDCD>,
+        delta: &NestedDelta<Self::ConceptId, SDCD, D>,
         concept: Self::ConceptId,
     ) -> Self::ConceptId {
         self.read_concept(delta, concept)
@@ -232,9 +243,11 @@ where
             })
             .unwrap_or(concept)
     }
-    fn is_disconnected(
+    fn is_disconnected<
+        D: SharedDelta<NestedDelta = NestedDelta<Self::ConceptId, SDCD, D>>,
+    >(
         &self,
-        delta: &ContextDelta<Self::ConceptId, SDCD>,
+        delta: &NestedDelta<Self::ConceptId, SDCD, D>,
         concept: Self::ConceptId,
     ) -> bool {
         self.read_concept(delta, concept).get_reduction().is_none()
@@ -251,9 +264,11 @@ where
                 .next()
                 .is_none()
     }
-    fn righthand_of_without_label_is_empty(
+    fn righthand_of_without_label_is_empty<
+        D: SharedDelta<NestedDelta = NestedDelta<Self::ConceptId, SDCD, D>>,
+    >(
         &self,
-        delta: &ContextDelta<Self::ConceptId, SDCD>,
+        delta: &NestedDelta<Self::ConceptId, SDCD, D>,
         con: Self::ConceptId,
     ) -> bool {
         self.concrete_concept_id(delta, ConcreteConceptType::Label)
@@ -263,18 +278,22 @@ where
             })
             .is_none()
     }
-    fn get_normal_form(
+    fn get_normal_form<
+        D: SharedDelta<NestedDelta = NestedDelta<Self::ConceptId, SDCD, D>>,
+    >(
         &self,
-        delta: &ContextDelta<Self::ConceptId, SDCD>,
+        delta: &NestedDelta<Self::ConceptId, SDCD, D>,
         concept: Self::ConceptId,
     ) -> Option<Self::ConceptId> {
         self.read_concept(delta, concept)
             .get_reduction()
             .map(|n| self.get_normal_form(delta, n).unwrap_or(n))
     }
-    fn get_concept_of_label(
+    fn get_concept_of_label<
+        D: SharedDelta<NestedDelta = NestedDelta<Self::ConceptId, SDCD, D>>,
+    >(
         &self,
-        delta: &ContextDelta<Self::ConceptId, SDCD>,
+        delta: &NestedDelta<Self::ConceptId, SDCD, D>,
         concept: Self::ConceptId,
     ) -> Option<Self::ConceptId> {
         let label_concept_id =
@@ -283,9 +302,11 @@ where
         self.read_concept(delta, concept)
             .find_as_hand_in_composition_with(label_concept_id, Hand::Right)
     }
-    fn contains(
+    fn contains<
+        D: SharedDelta<NestedDelta = NestedDelta<Self::ConceptId, SDCD, D>>,
+    >(
         &self,
-        delta: &ContextDelta<Self::ConceptId, SDCD>,
+        delta: &NestedDelta<Self::ConceptId, SDCD, D>,
         outer: Self::ConceptId,
         inner: Self::ConceptId,
     ) -> bool {
@@ -300,9 +321,11 @@ where
             false
         }
     }
-    fn check_reductions(
+    fn check_reductions<
+        D: SharedDelta<NestedDelta = NestedDelta<Self::ConceptId, SDCD, D>>,
+    >(
         &self,
-        delta: &ContextDelta<Self::ConceptId, SDCD>,
+        delta: &NestedDelta<Self::ConceptId, SDCD, D>,
         outer_concept: Self::ConceptId,
         inner_concept: Self::ConceptId,
     ) -> ZiaResult<()> {
@@ -318,9 +341,12 @@ where
             },
         )
     }
-    fn get_reduction_or_reduction_of_composition<'a>(
+    fn get_reduction_or_reduction_of_composition<
+        'a,
+        D: SharedDelta<NestedDelta = NestedDelta<Self::ConceptId, SDCD, D>>,
+    >(
         &'a self,
-        delta: &ContextDelta<Self::ConceptId, SDCD>,
+        delta: &NestedDelta<Self::ConceptId, SDCD, D>,
         concept: Self::ConceptId,
     ) -> Self::MixedConcept<'a> {
         self.read_concept(
@@ -330,14 +356,18 @@ where
             ),
         )
     }
-    fn concrete_concept_id(
+    fn concrete_concept_id<
+        D: SharedDelta<NestedDelta = NestedDelta<Self::ConceptId, SDCD, D>>,
+    >(
         &self,
-        delta: &ContextDelta<Self::ConceptId, SDCD>,
+        delta: &NestedDelta<Self::ConceptId, SDCD, D>,
         cc: ConcreteConceptType,
     ) -> Option<Self::ConceptId>;
-    fn concrete_concept_type(
+    fn concrete_concept_type<
+        D: SharedDelta<NestedDelta = NestedDelta<Self::ConceptId, SDCD, D>>,
+    >(
         &self,
-        delta: &ContextDelta<Self::ConceptId, SDCD>,
+        delta: &NestedDelta<Self::ConceptId, SDCD, D>,
         concept_id: Self::ConceptId,
     ) -> Option<ConcreteConceptType>;
 }
