@@ -42,20 +42,49 @@ pub fn view(model: &Model) -> impl IntoNodes<Msg> {
             },
             h1![C![C.flex_1], "Zia"],
             p![C![C.flex_1], "A programming language that defines itself"],
-            button![
-                C![
-                    C.flex_1,
-                    C.border_2,
-                    C.bg_primary,
-                    EDGE_STYLE,
-                    TEXT_PADDING,
-                    C.text_center,
-                    C.text_secondary
+            div![
+                C![C.flex, C.flex_2, C.flex_row],
+                p![
+                    C![
+                        C.flex,
+                        C.flex_1,
+                        C.text_center,
+                        C.justify_center,
+                        C.self_center
+                    ],
+                    "Tutorials"
                 ],
-                ev(Ev::Click, |_| Msg::Home(HomeMsg::StartTutorial(
-                    &TUTORIALS.0.steps
-                ))),
-                TUTORIALS.0.title
+                div![
+                    C![C.flex, C.flex_1, C.flex_col],
+                    button![
+                        C![
+                            C.flex_1,
+                            C.border_2,
+                            C.bg_primary,
+                            EDGE_STYLE,
+                            C.text_center,
+                            C.text_secondary
+                        ],
+                        ev(Ev::Click, |_| Msg::Home(HomeMsg::StartTutorial(
+                            &TUTORIALS.0.steps
+                        ))),
+                        TUTORIALS.0.title
+                    ],
+                    button![
+                        C![
+                            C.flex_1,
+                            C.border_2,
+                            C.bg_primary,
+                            EDGE_STYLE,
+                            C.text_center,
+                            C.text_secondary
+                        ],
+                        ev(Ev::Click, |_| Msg::Home(HomeMsg::StartTutorial(
+                            &TUTORIALS.1.steps
+                        ))),
+                        TUTORIALS.1.title
+                    ],
+                ]
             ],
             button![
                 C![
@@ -63,7 +92,6 @@ pub fn view(model: &Model) -> impl IntoNodes<Msg> {
                     C.border_2,
                     C.bg_primary,
                     EDGE_STYLE,
-                    TEXT_PADDING,
                     C.text_center,
                     C.text_secondary
                 ],
