@@ -1,7 +1,5 @@
 extern crate wasm_bindgen_test;
-use wasm_bindgen_test::*;
+use wasm_bindgen_test::wasm_bindgen_test;
 
 #[wasm_bindgen_test]
-fn pass() {
-    assert_eq!(1, 1);
-}
+const fn pass() {}
