@@ -1,14 +1,16 @@
-pub trait AndAlso<T> {
-    fn and_also<'a, 'b>(&'a self, other: &'b Self) -> Option<(&'a T, &'b T)>;
-    fn and_also_mut<'a, 'b>(
-        &'a mut self,
-        other: &'b mut Self,
-    ) -> Option<(&'a mut T, &'b mut T)>;
-    fn and_also_move(self, other: Self) -> Option<(T, T)>;
+pub trait AndAlso<T, U> {
+    fn and_also<'a, 'b>(
+        &'a self,
+        other: &'b Option<U>,
+    ) -> Option<(&'a T, &'b U)>;
+    fn and_also_move(self, other: Option<U>) -> Option<(T, U)>;
 }
 
-impl<T> AndAlso<T> for Option<T> {
-    fn and_also<'a, 'b>(&'a self, other: &'b Self) -> Option<(&'a T, &'b T)> {
+impl<T, U> AndAlso<T, U> for Option<T> {
+    fn and_also<'a, 'b>(
+        &'a self,
+        other: &'b Option<U>,
+    ) -> Option<(&'a T, &'b U)> {
         if let (Some(x), Some(y)) = (self, other) {
             Some((x, y))
         } else {
@@ -16,18 +18,7 @@ impl<T> AndAlso<T> for Option<T> {
         }
     }
 
-    fn and_also_mut<'a, 'b>(
-        &'a mut self,
-        other: &'b mut Self,
-    ) -> Option<(&'a mut T, &'b mut T)> {
-        if let (Some(x), Some(y)) = (self, other) {
-            Some((x, y))
-        } else {
-            None
-        }
-    }
-
-    fn and_also_move(self, other: Self) -> Option<(T, T)> {
+    fn and_also_move(self, other: Option<U>) -> Option<(T, U)> {
         if let (Some(x), Some(y)) = (self, other) {
             Some((x, y))
         } else {
