@@ -1,9 +1,15 @@
 use std::{
     borrow::Borrow,
-    collections::HashMap,
     fmt::{Debug, Display},
     hash::Hash,
     ops::Deref,
+};
+
+use maplit::hashmap;
+
+use crate::{
+    and_also::AndAlso, consistent_merge::ConsistentMerge,
+    substitute::Substitutions, variable_mask_list::VariableMask,
 };
 
 pub trait SyntaxTree
@@ -32,7 +38,7 @@ where
 
     fn is_leaf_variable(&self) -> bool;
 
-    fn new_constant_concept(concept_id: Self::ConceptId) -> Self;
+    fn new_constant_concept(concept_id: impl Into<Self::ConceptId>) -> Self;
 
     fn new_quantifier_concept(concept_id: Self::ConceptId) -> Self;
 
@@ -40,7 +46,10 @@ where
 
     fn new_leaf_variable(concept_id: Self::ConceptId) -> Self;
 
-    fn bind_nonquantifier_concept(self, concept: Self::ConceptId) -> Self;
+    fn bind_nonquantifier_concept(
+        self,
+        concept: impl Into<Self::ConceptId>,
+    ) -> Self;
 
     fn bind_nonquantifier_concept_as_ref(&mut self, concept: Self::ConceptId);
 
@@ -67,5 +76,34 @@ where
     fn check_example(
         example: &Self::SharedSyntax,
         generalisation: &Self::SharedSyntax,
-    ) -> Option<HashMap<Self::SharedSyntax, Self::SharedSyntax>>;
+    ) -> Option<ExampleSubstitutions<Self>>;
+}
+
+#[derive(Clone, Debug)]
+pub struct ExampleSubstitutions<S: SyntaxTree> {
+    pub generalisation: Substitutions<S::SharedSyntax>,
+    pub example: VariableMask<S>,
+}
+
+impl<S: SyntaxTree> Default for ExampleSubstitutions<S> {
+    fn default() -> Self {
+        Self {
+            generalisation: hashmap! {},
+            example: hashmap! {},
+        }
+    }
+}
+
+impl<S: SyntaxTree> ConsistentMerge for ExampleSubstitutions<S> {
+    type Output = Self;
+
+    fn consistent_merge(self, other: Self) -> Option<Self::Output> {
+        self.generalisation
+            .consistent_merge(other.generalisation)
+            .and_also_move(self.example.consistent_merge(other.example))
+            .map(|(generalisation, example)| Self {
+                generalisation,
+                example,
+            })
+    }
 }
