@@ -1,7 +1,7 @@
 use crate::{
     ast::SyntaxTree,
     concepts::ConceptTrait,
-    context_search::{ReductionReason, ReductionResult},
+    reduction_reason::{ReductionReason, ReductionResult},
 };
 use std::fmt::Debug;
 
@@ -45,5 +45,17 @@ where
         &self,
         ast: &SharedSyntax<Self::RR>,
         reduction_result: &ReductionResult<Self::RR>,
+    );
+
+    fn get_inference_or_else(
+        &self,
+        concept: ConceptId<Self::RR>,
+        infer: impl Fn() -> ReductionResult<Self::RR> + Copy,
+    ) -> ReductionResult<Self::RR>;
+
+    fn insert_inference(
+        &self,
+        concept: ConceptId<Self::RR>,
+        rr: &ReductionResult<Self::RR>,
     );
 }
