@@ -17,11 +17,20 @@
 use crate::{
     and_also::AndAlso,
     associativity::Associativity,
-    ast::SyntaxTree,
+    ast::{ExampleSubstitutions, SyntaxTree},
     concepts::{format_string, ConceptTrait, ConcreteConceptType, Hand},
+    consistent_merge::ConsistentMerge,
     context_cache::ContextCache,
-    context_delta::{ContextDelta, DirectConceptDelta},
+    context_delta::{
+        DirectConceptDelta, NestedDelta, NewConceptDelta, SharedDelta,
+    },
+    iteration::Iteration,
+    mixed_concept::MixedConcept,
+    reduction_reason::{
+        RRSharedSyntax, ReductionReason, ReductionResult, SharedSyntax, Syntax,
+    },
     snap_shot::Reader as SnapShotReader,
+    substitute::substitute,
     variable_mask_list::{VariableMask, VariableMaskList},
 };
 use log::debug;
@@ -29,439 +38,117 @@ use maplit::{hashmap, hashset};
 use std::{
     collections::{HashMap, HashSet},
     fmt::Debug,
+    marker::PhantomData,
 };
 
-pub type SharedSyntax<C> = <Syntax<C> as SyntaxTree>::SharedSyntax;
-pub type Syntax<C> = ReductionReasonSyntax<<C as ContextCache>::RR>;
 #[derive(Debug)]
-pub struct ContextSearch<'a, S, C, SDCD, VML>
+pub struct ContextSearch<'s, 'v, S, C, VML, SDCD, D, CCI: MixedConcept>
 where
-    S: SnapShotReader<SDCD>,
-    C: ContextCache,
-    SDCD: Clone
-        + AsRef<DirectConceptDelta<S::ConceptId>>
-        + From<DirectConceptDelta<S::ConceptId>>,
-    VML: VariableMaskList,
-{
-    snap_shot: &'a S,
-    variable_mask: VML::Shared,
-    delta: &'a ContextDelta<S::ConceptId, SDCD>,
-    caches: C,
-    syntax_evaluating: HashSet<SharedSyntax<C>>,
-    bound_variable_syntax: &'a HashSet<SharedSyntax<C>>,
-}
-
-pub type ReductionResult<RR> = Option<Reduction<RR>>;
-
-pub type Reduction<RR> = (ReductionReasonSharedSyntax<RR>, RR);
-
-macro_rules! impl_reduction_reason {
-    ($refcounter:tt, $reduction_reason:tt) => {
-        use crate::context_search::{
-            ComparisonReason, Reduction, ReductionReason, ReductionTruthResult,
-            Substitutions,
-        };
-        #[derive(Clone, PartialEq, Debug)]
-        pub enum $reduction_reason<S: SyntaxTree> {
-            Comparison($refcounter<ComparisonReason<Self>>),
-            Explicit,
-            Rule {
-                generalisation: S::SharedSyntax,
-                variable_mask: VariableMask<S>,
-                reason: $refcounter<Self>,
-            },
-            Inference {
-                implication: S::SharedSyntax,
-                reason: $refcounter<Self>,
-            },
-            Default {
-                operator: S::ConceptId,
-            },
-            Partial(HashMap<S::SharedSyntax, Reduction<Self>>),
-            Existence {
-                substitutions: Substitutions<S::SharedSyntax>,
-                generalisation: S::SharedSyntax,
-            },
-            Recursive {
-                syntax: S::SharedSyntax,
-                reason: $refcounter<Self>,
-                from: $refcounter<Self>,
-            },
-            SyntaxCannotReduceToItself,
-            LeftReducesToRight {
-                reason: $refcounter<Self>,
-                left: S::SharedSyntax,
-                right: S::SharedSyntax,
-            },
-            RightReducesToLeft {
-                reason: $refcounter<Self>,
-                left: S::SharedSyntax,
-                right: S::SharedSyntax,
-            },
-        }
-
-        impl<S: SyntaxTree> From<ComparisonReason<$reduction_reason<S>>>
-            for $reduction_reason<S>
-        {
-            fn from(
-                comparison_reason: ComparisonReason<$reduction_reason<S>>,
-            ) -> Self {
-                Self::Comparison(comparison_reason.into())
-            }
-        }
-
-        impl<S: SyntaxTree> $reduction_reason<S> {
-            fn determine_evidence_of_reduction(
-                left: &S::SharedSyntax,
-                right: &S::SharedSyntax,
-                reduce: impl Fn(&S::SharedSyntax) -> ReductionResult<Self>,
-            ) -> Option<Self> {
-                reduce(left).and_then(|(reduced_left, reason)| {
-                    if &reduced_left == right {
-                        Some(reason)
-                    } else {
-                        Self::determine_evidence_of_reduction(
-                            &reduced_left,
-                            right,
-                            reduce,
-                        )
-                        .map(|new_reason| Self::Recursive {
-                            syntax: left.clone(),
-                            reason: new_reason.into(),
-                            from: reason.into(),
-                        })
-                    }
-                })
-            }
-        }
-
-        impl<S: SyntaxTree> ReductionReason for $reduction_reason<S> {
-            type Syntax = S;
-
-            fn simplify_reasoning(
-                reason: Option<Self>,
-                reversed_reason: Option<Self>,
-            ) -> ComparisonReason<Self> {
-                match (&reason, &reversed_reason) {
-                    (Some(Self::Comparison(cr)), rr) => match cr.as_ref() {
-                        ComparisonReason::Reduction {
-                            reason: comparsion_reason,
-                            reversed_reason: reversed_comparison_reason,
-                        } if rr == reversed_comparison_reason => {
-                            ComparisonReason::Reduction {
-                                reversed_reason: reversed_comparison_reason
-                                    .clone(),
-                                reason: comparsion_reason.clone(),
-                            }
-                        },
-                        _ => ComparisonReason::Reduction {
-                            reason,
-                            reversed_reason,
-                        },
-                    },
-                    (r, Some(Self::Comparison(cr))) => match cr.as_ref() {
-                        ComparisonReason::Reduction {
-                            reason: reversed_comparison_reason,
-                            reversed_reason: comparison_reason,
-                        } if r == comparison_reason => {
-                            ComparisonReason::Reduction {
-                                reversed_reason: reversed_comparison_reason
-                                    .clone(),
-                                reason: comparison_reason.clone(),
-                            }
-                        },
-                        _ => ComparisonReason::Reduction {
-                            reason,
-                            reversed_reason,
-                        },
-                    },
-                    _ => ComparisonReason::Reduction {
-                        reason,
-                        reversed_reason,
-                    },
-                }
-            }
-
-            fn determine_reduction_truth(
-                left: &S::SharedSyntax,
-                right: &S::SharedSyntax,
-                reduce: impl Fn(&S::SharedSyntax) -> ReductionResult<Self>,
-            ) -> ReductionTruthResult<Self> {
-                if left == right {
-                    Some((false, Self::SyntaxCannotReduceToItself))
-                } else {
-                    match (
-                        Self::determine_evidence_of_reduction(
-                            left, right, &reduce,
-                        ),
-                        Self::determine_evidence_of_reduction(
-                            right, left, reduce,
-                        ),
-                    ) {
-                        (Some(_), Some(_)) => panic!(
-                            "{:#?} and {:#?} reduce to each other?!",
-                            left, right
-                        ),
-                        (Some(reason), None) => Some((
-                            true,
-                            Self::LeftReducesToRight {
-                                reason: reason.into(),
-                                left: left.clone(),
-                                right: right.clone(),
-                            },
-                        )),
-                        (None, Some(reason)) => Some((
-                            false,
-                            Self::RightReducesToLeft {
-                                reason: reason.into(),
-                                left: left.clone(),
-                                right: right.clone(),
-                            },
-                        )),
-                        (None, None) => None,
-                    }
-                }
-            }
-
-            fn recursive_reason(
-                previous: Option<Self>,
-                last: Self,
-                previously_reduced_syntax: &S::SharedSyntax,
-            ) -> Self {
-                previous.map_or(last.clone(), |from| Self::Recursive {
-                    from: from.into(),
-                    reason: last.into(),
-                    syntax: previously_reduced_syntax.clone(),
-                })
-            }
-
-            fn existence(
-                substitutions: Substitutions<S::SharedSyntax>,
-                generalisation: S::SharedSyntax,
-            ) -> Self {
-                Self::Existence {
-                    substitutions,
-                    generalisation,
-                }
-            }
-
-            fn inference(implication: S::SharedSyntax, reason: Self) -> Self {
-                Self::Inference {
-                    implication,
-                    reason: reason.into(),
-                }
-            }
-
-            fn explicit() -> Self {
-                Self::Explicit
-            }
-
-            fn default(operator: S::ConceptId) -> Self {
-                Self::Default {
-                    operator,
-                }
-            }
-
-            fn rule(
-                generalisation: S::SharedSyntax,
-                variable_mask: VariableMask<S>,
-                reason: Self,
-            ) -> Self {
-                Self::Rule {
-                    generalisation,
-                    variable_mask: variable_mask.clone(),
-                    reason: reason.into(),
-                }
-            }
-
-            fn partial(
-                partial_reductions: HashMap<S::SharedSyntax, Reduction<Self>>,
-            ) -> Self {
-                debug_assert!(partial_reductions.len() > 0);
-                let mut flattened_partial_reductions = hashmap! {};
-                for (unreduced_concept, (reduced_concept, reduction_reason)) in
-                    partial_reductions
-                {
-                    if let Self::Partial(inner_partial_reductions) =
-                        reduction_reason
-                    {
-                        flattened_partial_reductions
-                            .extend(inner_partial_reductions);
-                    } else {
-                        flattened_partial_reductions.insert(
-                            unreduced_concept,
-                            (reduced_concept, reduction_reason),
-                        );
-                    }
-                }
-                Self::Partial(flattened_partial_reductions)
-            }
-        }
-    };
-}
-
-pub trait ReductionReason
-where
-    Self: Clone + Debug + From<ComparisonReason<Self>> + PartialEq,
-{
-    type Syntax: SyntaxTree;
-
-    fn simplify_reasoning(
-        reason: Option<Self>,
-        reversed_reason: Option<Self>,
-    ) -> ComparisonReason<Self>;
-
-    fn determine_reduction_truth(
-        left: &ReductionReasonSharedSyntax<Self>,
-        right: &ReductionReasonSharedSyntax<Self>,
-        reduce: impl Fn(&ReductionReasonSharedSyntax<Self>) -> ReductionResult<Self>,
-    ) -> ReductionTruthResult<Self>;
-
-    fn recursive_reason(
-        previous: Option<Self>,
-        last: Self,
-        previously_reduced_syntax: &ReductionReasonSharedSyntax<Self>,
-    ) -> Self;
-
-    fn existence(
-        substitutions: Substitutions<ReductionReasonSharedSyntax<Self>>,
-        generalisation: ReductionReasonSharedSyntax<Self>,
-    ) -> Self;
-
-    fn inference(
-        implication: ReductionReasonSharedSyntax<Self>,
-        reason: Self,
-    ) -> Self;
-
-    fn explicit() -> Self;
-
-    fn default(operator_id: ReductionReasonConceptId<Self>) -> Self;
-
-    fn rule(
-        generalisation: ReductionReasonSharedSyntax<Self>,
-        variable_mask: VariableMask<Self::Syntax>,
-        reason: Self,
-    ) -> Self;
-
-    fn partial(
-        partial_reductions: HashMap<
-            ReductionReasonSharedSyntax<Self>,
-            Reduction<Self>,
-        >,
-    ) -> Self;
-}
-
-type ReductionReasonSharedSyntax<RR> =
-    <ReductionReasonSyntax<RR> as SyntaxTree>::SharedSyntax;
-type ReductionReasonConceptId<RR> =
-    <ReductionReasonSyntax<RR> as SyntaxTree>::ConceptId;
-type ReductionReasonSyntax<RR> = <RR as ReductionReason>::Syntax;
-
-pub type Substitutions<SharedSyntax> = HashMap<SharedSyntax, SharedSyntax>;
-
-fn substitute<Syntax: SyntaxTree>(
-    syntax: &mut Syntax,
-    substitutions: &Substitutions<Syntax::SharedSyntax>,
-) {
-    if let Some(substitution) = substitutions.get(syntax) {
-        *syntax = substitution.as_ref().clone();
-    } else if let Some((left, right)) = syntax.get_expansion_mut() {
-        substitute(left, substitutions);
-        substitute(right, substitutions);
-    }
-}
-
-impl<'a, S, C, SDCD, VML> ContextSearch<'a, S, C, SDCD, VML>
-where
-    S: SnapShotReader<SDCD> + Sync + std::fmt::Debug,
+    S: SnapShotReader<SDCD, ConceptId = CCI> + Sync + std::fmt::Debug,
     Self: Iteration<ConceptId = S::ConceptId, Syntax = Syntax<C>>,
     C: ContextCache,
     Syntax<C>: SyntaxTree<ConceptId = S::ConceptId>,
     SDCD: Clone
         + AsRef<DirectConceptDelta<S::ConceptId>>
-        + From<DirectConceptDelta<S::ConceptId>>,
+        + From<DirectConceptDelta<S::ConceptId>>
+        + Debug,
     VML: VariableMaskList<Syntax = Syntax<C>>,
+    D: SharedDelta<NestedDelta = NestedDelta<S::ConceptId, SDCD, D>>,
 {
-    fn infer_reduction<'b>(
-        &'b self,
-        concept: &S::MixedConcept<'b>,
+    snap_shot: &'s S,
+    variable_mask: VML::Shared,
+    delta: D,
+    caches: C,
+    syntax_evaluating: HashSet<SharedSyntax<C>>,
+    concept_inferring: HashSet<S::ConceptId>,
+    bound_variable_syntax: &'v HashSet<SharedSyntax<C>>,
+    phantom: PhantomData<SDCD>,
+    phantom2: PhantomData<CCI>,
+}
+
+impl<'s, 'v, S, C, SDCD, VML, D, CCI: MixedConcept>
+    ContextSearch<'s, 'v, S, C, VML, SDCD, D, CCI>
+where
+    S: SnapShotReader<SDCD, ConceptId = CCI> + Sync + std::fmt::Debug,
+    Self: Iteration<ConceptId = S::ConceptId, Syntax = Syntax<C>>,
+    C: ContextCache,
+    Syntax<C>: SyntaxTree<ConceptId = S::ConceptId>,
+    SDCD: Clone
+        + AsRef<DirectConceptDelta<S::ConceptId>>
+        + From<DirectConceptDelta<S::ConceptId>>
+        + Debug,
+    VML: VariableMaskList<Syntax = Syntax<C>>,
+    D: AsRef<NestedDelta<S::ConceptId, SDCD, D>>
+        + SharedDelta<NestedDelta = NestedDelta<S::ConceptId, SDCD, D>>,
+{
+    fn infer_reduction(
+        &self,
+        concept: &S::MixedConcept<'_>,
     ) -> ReductionResult<C::RR> {
         debug!("infer_reduction({:#?})", concept);
         let implication_id =
             self.concrete_concept_id(ConcreteConceptType::Implication)?;
-        let composition_id = concept
-            .find_as_hand_in_composition_with(implication_id, Hand::Right)?;
-        let composition_concept =
-            self.snap_shot.read_concept(self.delta, composition_id);
-        let result = composition_concept.iter_hand_of(Hand::Right).find_map(
-            |(condition_id, implication_rule_id)| {
-                let condition = self.to_ast(condition_id);
-                let (reduced_condition, reason) = self.reduce(&condition)?;
-                let x = reduced_condition.get_concept()?;
-                self.is_concrete_type(ConcreteConceptType::True, x).map(|x| {
-                    (
-                        self.to_ast(x),
-                        C::RR::inference(
-                            self.to_ast(implication_rule_id),
-                            reason,
-                        ),
-                    )
-                })
-            },
-        );
-        result
-    }
-
-    fn is_concrete_type(
-        &self,
-        cct: ConcreteConceptType,
-        concept_id: S::ConceptId,
-    ) -> Option<S::ConceptId> {
-        if Some(cct) == self.concrete_type(concept_id) {
-            Some(concept_id)
-        } else {
-            None
-        }
-    }
-
-    fn concrete_type(
-        &self,
-        concept_id: S::ConceptId,
-    ) -> Option<ConcreteConceptType> {
-        self.snap_shot.concrete_concept_type(self.delta, concept_id)
-    }
-
-    fn concrete_type_of_ast(
-        &self,
-        ast: &SharedSyntax<C>,
-    ) -> Option<ConcreteConceptType> {
-        ast.get_concept().and_then(|c| self.concrete_type(c))
-    }
-
-    fn concrete_concept_id(
-        &self,
-        cct: ConcreteConceptType,
-    ) -> Option<S::ConceptId> {
-        self.snap_shot.concrete_concept_id(self.delta, cct)
+        let result = concept
+            .find_as_hand_in_composition_with(implication_id, Hand::Right)
+            .and_then(|composition_id| {
+                let composition_concept = self
+                    .snap_shot
+                    .read_concept(self.delta.as_ref(), composition_id);
+                // TODO: check if implication is actually true
+                // TODO: check for any concepts that are the reduction of `concept`
+                let result = composition_concept
+                    .iter_hand_of(Hand::Right)
+                    .find_map(|(condition_id, implication_rule_id)| {
+                        let condition = self.to_ast(&condition_id);
+                        let (reduced_condition, reason) =
+                            self.reduce(&condition)?;
+                        let x = reduced_condition.get_concept()?;
+                        self.is_concrete_type(ConcreteConceptType::True, &x)
+                            .map(|x| {
+                                (
+                                    // TODO: this should be the "true" concept only if the implication result is equivalent to the `concept`
+                                    self.to_ast(&x),
+                                    C::RR::inference(
+                                        self.to_ast(&implication_rule_id),
+                                        reason,
+                                    ),
+                                )
+                            })
+                    });
+                result
+            });
+        result.or_else(|| {
+            self.caches.get_inference_or_else(concept.id(), || {
+                let ast = &self.to_ast(&concept.id());
+                let rr = self.find_examples_of_inferred_reduction(ast);
+                self.caches.insert_inference(concept.id(), &rr);
+                rr
+            })
+        })
     }
 
     pub fn concrete_ast(
         &self,
         cct: ConcreteConceptType,
     ) -> Option<SharedSyntax<C>> {
-        self.concrete_concept_id(cct).map(|id| self.to_ast(id))
+        self.concrete_concept_id(cct).map(|id| self.to_ast(&id))
     }
 
     /// Returns the syntax for the reduction of a concept.
-    fn reduce_concept(&self, id: S::ConceptId) -> ReductionResult<C::RR> {
+    fn reduce_concept(&self, id: &S::ConceptId) -> ReductionResult<C::RR> {
         debug!("reduce_concept({})", id);
-        let concept = self.snap_shot.read_concept(self.delta, id);
-        self.infer_reduction(&concept).or_else(|| {
+        let concept = self.snap_shot.read_concept(self.delta.as_ref(), *id);
+        if self.concept_inferring.contains(id) {
+            None
+        } else {
+            self.infer_reduction(&concept)
+        }
+        .or_else(|| {
             let n = concept.get_reduction()?;
-            if self.is_leaf_variable(n) {
+            if self.is_leaf_variable(&n) {
                 self.variable_mask.get(n).cloned()
             } else {
-                Some(self.to_ast(n))
+                Some(self.to_ast(&n))
             }
             .map(|r| (r, C::RR::explicit()))
         })
@@ -473,17 +160,17 @@ where
         ast: &SharedSyntax<C>,
         left: &SharedSyntax<C>,
         right: &SharedSyntax<C>,
-        operator_id: S::ConceptId,
-        default_concept_id: S::ConceptId,
+        operator_id: &S::ConceptId,
+        default_concept_id: &S::ConceptId,
     ) -> ReductionResult<C::RR> {
         let mut reduced_pair: ReductionResult<C::RR> = None;
         let mut operator_composition_check = || {
             let cache = <C as ContextCache>::SharedReductionCache::default();
-            let mut context_search = self.spawn(&cache);
+            let mut context_search = self.spawn(&cache, self.delta.clone());
             context_search.syntax_evaluating.insert(ast.clone());
             reduced_pair = context_search.reduce_pair(left, right);
             let operator_concept =
-                self.snap_shot.read_concept(self.delta, operator_id);
+                self.snap_shot.read_concept(self.delta.as_ref(), *operator_id);
             let find = |c| {
                 operator_concept.find_as_hand_in_composition_with(c, Hand::Left)
             };
@@ -492,7 +179,10 @@ where
         };
         if self.syntax_evaluating.contains(ast) || operator_composition_check()
         {
-            Some((self.to_ast(default_concept_id), C::RR::default(operator_id)))
+            Some((
+                self.to_ast(default_concept_id),
+                C::RR::default(*operator_id),
+            ))
         } else {
             reduced_pair
         }
@@ -502,13 +192,18 @@ where
     pub fn reduce(&self, ast: &SharedSyntax<C>) -> ReductionResult<C::RR> {
         debug!("reduce({})", ast.to_string());
         self.caches.get_reduction_or_else(ast, || {
-            let reduction_result = ast
-                .get_concept()
-                .and_then(|c| self.reduce_concept(c))
+            let maybe_concept: Option<CCI> = ast.get_concept();
+            if let Some(id) = maybe_concept {
+                if self.concrete_type(&id).is_some() {
+                    return None;
+                }
+            }
+            let reduction_result = maybe_concept
+                .and_then(|c| self.reduce_concept(&c))
                 .or_else(|| {
                     let (ref left, ref right) = ast.get_expansion()?;
                     left.get_concept()
-                        .and_then(|lc| match self.concrete_type(lc) {
+                        .and_then(|lc| match self.concrete_type(&lc) {
                             Some(ConcreteConceptType::Precedence) => {
                                 let default_concept_id = self
                                     .concrete_concept_id(
@@ -518,8 +213,8 @@ where
                                     ast,
                                     left,
                                     right,
-                                    lc,
-                                    default_concept_id,
+                                    &lc,
+                                    &default_concept_id,
                                 )
                             },
                             Some(ConcreteConceptType::Associativity) => {
@@ -531,8 +226,8 @@ where
                                     ast,
                                     left,
                                     right,
-                                    lc,
-                                    default_concept_id,
+                                    &lc,
+                                    &default_concept_id,
                                 )
                             },
                             _ => None,
@@ -585,17 +280,18 @@ where
                 .filter_generalisations_for_pair(left, right)
                 .iter()
                 .find_map(|(generalisation, variable_mask)| {
-                    let mut context_search = self.spawn(&cache);
+                    let mut context_search =
+                        self.spawn(&cache, self.delta.clone());
                     // Stack overflow occurs if you remove this
                     context_search
                         .insert_variable_mask(variable_mask.clone())
                         .ok()?;
-                    context_search.reduce_concept(*generalisation).map(
+                    context_search.reduce_concept(generalisation).map(
                         |(ast, reason)| {
                             (
                                 context_search.substitute(&ast, variable_mask),
                                 C::RR::rule(
-                                    self.to_ast(*generalisation),
+                                    self.to_ast(generalisation),
                                     variable_mask.clone(),
                                     reason,
                                 ),
@@ -683,79 +379,27 @@ where
             .get_concept()
             .and_also(&righthand.get_concept())
             .and_then(|(lc, rc)| {
-                let left_concept = self.snap_shot.read_concept(self.delta, *lc);
+                let left_concept =
+                    self.snap_shot.read_concept(self.delta.as_ref(), *lc);
                 left_concept
                     .find_as_hand_in_composition_with(*rc, Hand::Left)
                     .map(|def| {
                         let syntax = Syntax::<C>::from(
                             self.snap_shot
-                                .get_label(self.delta, def)
+                                .get_label(self.delta.as_ref(), def)
                                 .unwrap_or_else(|| {
                                     self.display_joint(lefthand, righthand)
                                 }),
                         );
-                        self.snap_shot
-                            .bind_concept_to_syntax(self.delta, syntax, def)
+                        self.snap_shot.bind_concept_to_syntax(
+                            self.delta.as_ref(),
+                            syntax,
+                            def,
+                        )
                     })
             })
             .unwrap_or_else(|| self.display_joint(lefthand, righthand).into())
             .bind_pair(lefthand.clone(), righthand.clone())
-    }
-
-    /// Checks if the concepts in the composition of `generalisation` matches the corresponding nodes in `ast` or is a free variable
-    /// returns `None` if does not match otherwise returns a `HashMap` mapping free variable concept IDs to the matching nodes in `ast`
-    pub fn check_generalisation(
-        &self,
-        ast: &SharedSyntax<C>,
-        generalisation: S::ConceptId,
-    ) -> Option<VariableMask<Syntax<C>>> {
-        (self.is_free_variable(generalisation)
-            && !self.bound_variable_syntax.contains(ast)
-            && !ast.get_concept().map_or(false, |c| {
-                self.snap_shot.read_concept(self.delta, c).bounded_variable()
-            }))
-        .then(|| {
-            if let Some((gl, gr)) = self
-                .snap_shot
-                .read_concept(self.delta, generalisation)
-                .get_composition()
-            {
-                let (l, r) = ast.get_expansion()?;
-                match (self.is_free_variable(gl), self.is_free_variable(gr)) {
-                    (true, true) => self
-                        .check_generalisation(&l, gl)
-                        .and_also_move(self.check_generalisation(&r, gr))
-                        .and_then(|(lm, mut rm)| {
-                            for (lmk, lmv) in lm {
-                                if let Some(rmv) = rm.get(&lmk) {
-                                    if rmv != &lmv {
-                                        return None;
-                                    }
-                                } else {
-                                    rm.insert(lmk, lmv);
-                                }
-                            }
-                            Some(rm)
-                        }),
-                    (true, false) if r.get_concept() == Some(gr) => {
-                        self.check_generalisation(&l, gl)
-                    },
-                    (false, true) if l.get_concept() == Some(gl) => {
-                        self.check_generalisation(&r, gr)
-                    },
-                    (false, false)
-                        if l.get_concept() == Some(gl)
-                            && r.get_concept() == Some(gr) =>
-                    {
-                        Some(hashmap! {})
-                    },
-                    _ => None,
-                }
-            } else {
-                Some(hashmap! {generalisation => ast.clone()})
-            }
-        })
-        .flatten()
     }
 
     fn find_generalisations(&self, ast: &Syntax<C>) -> HashSet<S::ConceptId> {
@@ -764,49 +408,33 @@ where
             if let Some(c) = l.get_concept() {
                 generalisations.extend(
                     self.snap_shot
-                        .read_concept(self.delta, c)
+                        .read_concept(self.delta.as_ref(), c)
                         .iter_composition_ids(Hand::Left),
                 );
             }
             if let Some(c) = r.get_concept() {
                 generalisations.extend(
                     self.snap_shot
-                        .read_concept(self.delta, c)
+                        .read_concept(self.delta.as_ref(), c)
                         .iter_composition_ids(Hand::Right),
                 );
             }
             self.find_generalisations(&l).iter().for_each(|g| {
                 generalisations.extend(
                     self.snap_shot
-                        .read_concept(self.delta, *g)
+                        .read_concept(self.delta.as_ref(), *g)
                         .iter_composition_ids(Hand::Left),
                 );
             });
             self.find_generalisations(&r).iter().for_each(|g| {
                 generalisations.extend(
                     self.snap_shot
-                        .read_concept(self.delta, *g)
+                        .read_concept(self.delta.as_ref(), *g)
                         .iter_composition_ids(Hand::Right),
                 );
             });
         }
         generalisations
-    }
-
-    fn is_leaf_variable(&self, lv: S::ConceptId) -> bool {
-        self.is_free_variable(lv) && self.is_leaf_concept(lv)
-    }
-
-    fn is_free_variable(&self, v: S::ConceptId) -> bool {
-        self.snap_shot.read_concept(self.delta, v).free_variable()
-            && self
-                .variable_mask
-                .tail()
-                .map_or(true, |vml| vml.get(v).is_none()) // A hack required to prevent stack overflow
-    }
-
-    fn is_leaf_concept(&self, l: S::ConceptId) -> bool {
-        self.snap_shot.read_concept(self.delta, l).get_composition().is_none()
     }
 
     /// Reduces the syntax as much as possible (returns the normal form syntax).
@@ -843,14 +471,18 @@ where
                     .concrete_concept_id(ConcreteConceptType::True)
                     .expect("true concept must exist");
                 let true_concept =
-                    self.snap_shot.read_concept(self.delta, true_id);
+                    self.snap_shot.read_concept(self.delta.as_ref(), true_id);
                 let truths = true_concept.find_what_reduces_to_it();
                 self.find_example(right, &truths.collect()).map(
                     |substitutions| {
-                        let true_syntax = self.to_ast(true_id);
+                        // TODO: determine whether substitutions.example should be considered
+                        let true_syntax = self.to_ast(&true_id);
                         (
                             true_syntax,
-                            C::RR::existence(substitutions, right.clone()),
+                            C::RR::existence(
+                                substitutions.generalisation,
+                                right.clone(),
+                            ),
                         )
                     },
                 )
@@ -863,160 +495,301 @@ where
         &self,
         generalisation: &SharedSyntax<C>,
         truths: &HashSet<S::ConceptId>,
-    ) -> Option<Substitutions<SharedSyntax<C>>> {
+    ) -> Option<ExampleSubstitutions<Syntax<C>>> {
         self.find_examples(generalisation, truths).pop()
+    }
+
+    pub fn find_examples_of_inferred_reduction(
+        &self,
+        ast_to_reduce: &SharedSyntax<C>,
+    ) -> ReductionResult<C::RR> {
+        let implication_id =
+            self.concrete_concept_id(ConcreteConceptType::Implication)?;
+        let reduction_operator =
+            self.concrete_ast(ConcreteConceptType::Reduction)?;
+        let mut spawned_delta = NestedDelta::spawn(self.delta.clone());
+        let variable_reduction_id = spawned_delta
+            .insert_delta_for_new_concept(NewConceptDelta::BoundVariable);
+        let variable_reduction_syntax =
+            Syntax::<C>::new_leaf_variable(variable_reduction_id).share();
+        let variable_condition_id = spawned_delta
+            .insert_delta_for_new_concept(NewConceptDelta::BoundVariable);
+        let variable_condition_syntax =
+            Syntax::<C>::new_leaf_variable(variable_condition_id).share();
+        let variable_result_id = spawned_delta
+            .insert_delta_for_new_concept(NewConceptDelta::BoundVariable);
+        let _variable_result_syntax =
+            Syntax::<C>::new_leaf_variable(variable_result_id).share();
+        let cache = <C as ContextCache>::SharedReductionCache::default();
+        let mut spawned_context_search =
+            self.spawn(&cache, D::from_nested(spawned_delta));
+        if let Some(concept) = ast_to_reduce.get_concept() {
+            spawned_context_search.concept_inferring.insert(concept);
+        }
+        let implication_syntax = spawned_context_search.to_ast(&implication_id);
+        let implication_rule_fn =
+            |condition: &SharedSyntax<C>,
+             prereduction: &SharedSyntax<C>,
+             reduction: &SharedSyntax<C>| {
+                Syntax::<C>::new_pair(
+                    condition.clone(),
+                    Syntax::<C>::new_pair(
+                        implication_syntax.clone(),
+                        Syntax::<C>::new_pair(
+                            prereduction.clone(),
+                            Syntax::<C>::new_pair(
+                                reduction_operator.clone(),
+                                reduction.clone(),
+                            )
+                            .share(),
+                        )
+                        .share(),
+                    )
+                    .share(),
+                )
+            };
+        let implication_rule_pattern = implication_rule_fn(
+            &variable_condition_syntax,
+            ast_to_reduce,
+            &variable_reduction_syntax,
+        );
+        let true_id = spawned_context_search
+            .concrete_concept_id(ConcreteConceptType::True)
+            .expect("true concept must exist");
+        let truths = self.equivalent_concepts_to(&true_id);
+        let irp = implication_rule_pattern.share();
+        spawned_context_search
+            .find_examples(&irp, &truths)
+            .into_iter()
+            .find_map(|substitutions| {
+                substitutions
+                    .generalisation
+                    .get(&variable_condition_syntax)
+                    .and_then(|condition_syntax| {
+                        let substituted_condition = self.substitute(
+                            condition_syntax,
+                            &substitutions.example,
+                        );
+                        let (condition_normal_form, reason) =
+                            spawned_context_search
+                                .recursively_reduce(&substituted_condition);
+                        spawned_context_search.is_concrete_type(
+                            ConcreteConceptType::True,
+                            &condition_normal_form.get_concept()?,
+                        )?;
+                        substitutions
+                            .generalisation
+                            .get(&variable_reduction_syntax)
+                            .map(|result| {
+                                (
+                                    self.substitute(
+                                        result,
+                                        &substitutions.example,
+                                    ),
+                                    C::RR::inference(
+                                        implication_rule_fn(
+                                            condition_syntax,
+                                            ast_to_reduce,
+                                            result,
+                                        )
+                                        .share(),
+                                        reason.unwrap_or_else(
+                                            ReductionReason::explicit,
+                                        ),
+                                    ),
+                                )
+                            })
+                    })
+            })
     }
 
     // TODO Lazily compute the concepts that are equivalent to a given normal form
     // until a required number of examples are found
+    // `generalisation` needs to contain bound variables
     fn find_examples(
         &self,
-        // needs to contain bounded variables
         generalisation: &SharedSyntax<C>,
         equivalence_set: &HashSet<S::ConceptId>, /* All concepts that are equal to generalisation */
-    ) -> Vec<Substitutions<SharedSyntax<C>>> {
+    ) -> Vec<ExampleSubstitutions<Syntax<C>>> {
         if let Some((left, right)) = generalisation.get_expansion() {
-            match (
-                self.contains_bound_variable_syntax(&left),
-                self.contains_bound_variable_syntax(&right),
-            ) {
-                (true, true) => {
-                    equivalence_set.iter().filter_map(|equivalent_concept_id| {
-                        let equivalent_concept = self
-                            .snap_shot
-                            .read_concept(self.delta, *equivalent_concept_id);
-                        equivalent_concept.get_composition()
-                    }).flat_map(|(equivalent_left_id, equivalent_right_id)| {
-                        let equivalent_left = self
-                            .snap_shot
-                            .read_concept(self.delta, equivalent_left_id);
-                        // TODO handle case when a concept implicitly reduces to `equivalent_left`
-                        let mut equivalent_left_equivalence_set: HashSet<S::ConceptId> =
-                            equivalent_left
-                                .find_what_reduces_to_it()
-                                .collect();
-                        equivalent_left_equivalence_set.insert(equivalent_left_id);
-                        let left_examples = self.find_examples(
-                            &left,
-                            &equivalent_left_equivalence_set,
-                        );
-                        let equivalent_right = self
-                            .snap_shot
-                            .read_concept(self.delta, equivalent_right_id);
-                        // TODO handle case when a concept implicitly reduces to `equivalent_right`
-                        let mut equivalent_right_equivalence_set: HashSet<
-                            S::ConceptId,
-                        > = equivalent_right
-                            .find_what_reduces_to_it()
-                            .collect();
-                        equivalent_right_equivalence_set
-                            .insert(equivalent_right_id);
-                        let right_examples = self.find_examples(
-                            &right,
-                            &equivalent_right_equivalence_set,
-                        );
-                        left_examples.into_iter().flat_map(|left_example| {
-                            let mut right_clone = right.clone();
-                            let mutable_right = Syntax::<C>::make_mut(&mut right_clone);
-                            substitute::<Syntax<C>>(mutable_right, &left_example);
-                            if self.contains_bound_variable_syntax(&right_clone) {
-                                self.find_examples(&right_clone, &equivalent_right_equivalence_set).into_iter().map(|mut right_example| {
-                                    right_example.extend(left_example.iter().map(|(k, v)| (k.clone(), v.clone())));
-                                    right_example
-                                }).collect::<Vec<_>>()
-                            } else if self.recursively_reduce(&right_clone).0.get_concept().map_or(false, |id| equivalent_right_equivalence_set.contains(&id)) {
-                                vec![left_example]
-                            } else {
-                                vec![]
-                            }
-                        }).chain(right_examples.into_iter().flat_map(|right_example| {
-                            let mut left_clone = left.clone();
-                            let mutable_left = Syntax::<C>::make_mut(&mut left_clone);
-                            substitute::<Syntax<C>>(mutable_left, &right_example);
-                            if self.contains_bound_variable_syntax(&left_clone) {
-                                self.find_examples(&left_clone, &equivalent_left_equivalence_set).into_iter().map(|mut left_example| {
-                                    left_example.extend(right_example.iter().map(|(k, v)| (k.clone(), v.clone())));
-                                    left_example
-                                }).collect::<Vec<_>>()
-                            } else if self.recursively_reduce(&left_clone).0.get_concept().map_or(false, |id| equivalent_left_equivalence_set.contains(&id)) {
-                                vec![right_example]
-                            } else {
-                                vec![]
-                            }
-                        })).collect::<Vec<_>>()
-                    }).collect()
-                },
-                (true, false) => self.find_examples_of_half_generalisation(
-                    &left,
-                    &right,
-                    equivalence_set,
-                    Hand::Right,
-                ),
-                (false, true) => self.find_examples_of_half_generalisation(
-                    &right,
-                    &left,
-                    equivalence_set,
-                    Hand::Left,
-                ),
-                (false, false) => vec![],
-            }
+            self.find_examples_of_branched_generalisation(
+                &left,
+                &right,
+                equivalence_set,
+            )
         } else {
-            debug_assert!(self.contains_bound_variable_syntax(generalisation));
+            debug_assert!(
+                self.contains_bound_variable_syntax(generalisation),
+                "Generalisation ({generalisation}) doesn't contain bound variables"
+            );
             equivalence_set
-                .iter()
+            .iter()
                 .map(|c| {
-                    let example = self.to_ast(*c);
-                    hashmap! {generalisation.clone() => example}
+                    let example = self.to_ast(c);
+                    ExampleSubstitutions {
+                        generalisation: hashmap! {generalisation.clone() => example},
+                        example: hashmap!{}
+                    }
                 })
                 .collect()
         }
     }
 
-    fn find_examples_of_half_generalisation(
+    fn find_examples_of_branched_generalisation(
         &self,
-        generalisated_part: &SharedSyntax<C>,
-        non_generalised_part: &SharedSyntax<C>,
-        equivalence_set_of_composition: &HashSet<S::ConceptId>,
-        non_generalised_hand: Hand,
-    ) -> Vec<Substitutions<SharedSyntax<C>>> {
-        non_generalised_part.get_concept().map_or_else(
-            Vec::new,
-            |non_generalised_id| {
-                let non_generalised_concept =
-                    self.snap_shot.read_concept(self.delta, non_generalised_id);
-                let examples =
-                    non_generalised_concept.iter_hand_of(non_generalised_hand);
-                examples
-                    .filter_map(|(generalised_hand, composition)| {
-                        equivalence_set_of_composition
-                            .contains(&composition)
-                            .then_some(generalised_hand)
+        left: &SharedSyntax<C>,
+        right: &SharedSyntax<C>,
+        equivalence_set: &HashSet<S::ConceptId>,
+    ) -> Vec<ExampleSubstitutions<Syntax<C>>> {
+        match (
+            self.contains_bound_variable_syntax(left),
+            self.contains_bound_variable_syntax(right),
+        ) {
+            (true, true) => {
+                equivalence_set
+                    .iter()
+                    .filter_map(|equivalent_concept_id| {
+                        self.composition_of_concept(equivalent_concept_id)
                     })
-                    .filter_map(|generalised_hand| {
-                        Syntax::<C>::check_example(
-                            &self.to_ast(generalised_hand),
-                            generalisated_part,
+                    .flat_map(|(equivalent_left_id, equivalent_right_id)| {
+                        let equivalent_left_equivalence_set =
+                            self.equivalent_concepts_to(&equivalent_left_id);
+                        let equivalent_right_equivalence_set =
+                            self.equivalent_concepts_to(&equivalent_right_id);
+                        self.find_examples(
+                            left,
+                            &equivalent_left_equivalence_set,
                         )
-                        .or_else(|| {
-                            // TODO handle case when a concept implicitly reduces to `non_generalised_hand`
-                            let mut equivalence_set =
-                                hashset! {generalised_hand};
-                            let non_generalised_hand_concept = self
-                                .snap_shot
-                                .read_concept(self.delta, generalised_hand);
-                            equivalence_set.extend(
-                                non_generalised_hand_concept
-                                    .find_what_reduces_to_it(),
+                        .into_iter()
+                        .flat_map(|left_example| {
+                            let mut right_clone = right.clone();
+                            let mutable_right =
+                                Syntax::<C>::make_mut(&mut right_clone);
+                            substitute::<Syntax<C>>(
+                                mutable_right,
+                                &left_example.generalisation,
                             );
-                            self.find_example(
-                                generalisated_part,
-                                &equivalence_set,
-                            )
+                            // `right_clone` now has any generalised variables matched in the example of `left` substituted for
+                            // but may still have some unmatched generalised variables
+                            let substituted_right = self.substitute(
+                                &right_clone,
+                                &left_example.example,
+                            );
+                            if self.contains_bound_variable_syntax(
+                                &substituted_right,
+                            ) {
+                                self.find_examples(
+                                    &substituted_right,
+                                    &equivalent_right_equivalence_set,
+                                )
+                                .into_iter()
+                                .filter_map(|right_example| {
+                                    right_example
+                                        .consistent_merge(left_example.clone())
+                                })
+                                .collect::<Vec<_>>()
+                            } else if self
+                                .recursively_reduce(&substituted_right)
+                                .0
+                                .get_concept()
+                                .map_or(false, |id| {
+                                    equivalent_right_equivalence_set
+                                        .contains(&id)
+                                })
+                            {
+                                vec![left_example]
+                            } else {
+                                vec![]
+                            }
                         })
+                        .collect::<Vec<_>>()
                     })
                     .collect()
             },
-        )
+            (true, false) => self.find_examples_of_half_generalisation(
+                left,
+                right,
+                equivalence_set,
+                Hand::Right,
+            ),
+            (false, true) => self.find_examples_of_half_generalisation(
+                right,
+                left,
+                equivalence_set,
+                Hand::Left,
+            ),
+            (false, false) => vec![],
+        }
+    }
+
+    fn find_examples_of_half_generalisation(
+        &self,
+        generalised_part: &SharedSyntax<C>,
+        non_generalised_part: &SharedSyntax<C>,
+        equivalence_set_of_composition: &HashSet<S::ConceptId>,
+        non_generalised_hand: Hand,
+    ) -> Vec<ExampleSubstitutions<Syntax<C>>> {
+        let mut subs = Vec::new();
+        subs.extend(equivalence_set_of_composition.iter().find_map(|equivalent_concept_id| {
+            let equivalent_concept = self
+                .snap_shot
+                .read_concept(self.delta.as_ref(), *equivalent_concept_id);
+            let (left, right) = equivalent_concept.get_composition()?;
+            let (equivalent_non_generalised_hand, equivalent_generalised_hand) = match non_generalised_hand {
+                Hand::Left => (left, right),
+                Hand::Right => (right, left)
+            };
+            if Some(equivalent_non_generalised_hand) != non_generalised_part.get_concept() {
+                if self.snap_shot.read_concept(self.delta.as_ref(), equivalent_non_generalised_hand).free_variable() {
+                    return self.find_example(generalised_part, &hashset!{equivalent_generalised_hand}).and_then(|subs| {
+                        // Could have a more efficient method for this
+                        subs.consistent_merge(ExampleSubstitutions{example: hashmap!{equivalent_non_generalised_hand => non_generalised_part.clone()}, ..Default::default()})
+                    })
+                }
+                return None;
+            }
+            self.find_example(generalised_part, &hashset!{equivalent_generalised_hand})
+        }));
+
+        let Some(non_generalised_id) = non_generalised_part.get_concept()
+        else {
+            return subs;
+        };
+        let non_generalised_concept = self
+            .snap_shot
+            .read_concept(self.delta.as_ref(), non_generalised_id);
+        let examples =
+            non_generalised_concept.iter_hand_of(non_generalised_hand);
+        subs.extend(
+            examples
+                .flat_map(|(example_hand, composition)| {
+                    let mut generalised_hands_and_substitutions = vec![];
+                    if equivalence_set_of_composition.contains(&composition) {
+                        generalised_hands_and_substitutions.push(example_hand);
+                    }
+                    generalised_hands_and_substitutions
+                })
+                .filter_map(|example_hand| {
+                    let example_hand_syntax = self.to_ast(&example_hand);
+                    Syntax::<C>::check_example(
+                        &example_hand_syntax,
+                        generalised_part,
+                    )
+                    .or_else(|| {
+                        // TODO handle case when a concept implicitly reduces to `non_generalised_hand`
+                        let mut equivalence_set = hashset! {example_hand};
+                        let non_generalised_hand_concept = self
+                            .snap_shot
+                            .read_concept(self.delta.as_ref(), example_hand);
+                        equivalence_set.extend(
+                            non_generalised_hand_concept
+                                .find_what_reduces_to_it(),
+                        );
+                        self.find_example(generalised_part, &equivalence_set)
+                    })
+                }),
+        );
+        subs
     }
 
     // Reduces a syntax tree based on the properties of the left branch and the branches of the right branch
@@ -1063,28 +836,32 @@ where
     }
 
     /// Returns the syntax for a concept. Panics if there is no concept with the given `concept_id`
-    pub fn to_ast(&self, concept_id: S::ConceptId) -> SharedSyntax<C> {
+    pub fn to_ast(
+        &self,
+        concept_id: &(impl Into<S::ConceptId> + Copy),
+    ) -> SharedSyntax<C> {
+        let concept_id = (*concept_id).into();
         self.variable_mask.get(concept_id).cloned().unwrap_or_else(|| {
             self.caches.get_syntax_tree_or_else(concept_id, || {
                 let concept =
-                    self.snap_shot.read_concept(self.delta, concept_id);
+                self.snap_shot.read_concept(self.delta.as_ref(), concept_id);
                 let syntax = concept.get_string().map_or_else(
-                    || self.snap_shot.get_label(self.delta, concept_id),
+                    || self.snap_shot.get_label(self.delta.as_ref(), concept_id),
                     |s| Some(format_string(&s)),
                 ).map_or_else(|| {
                     if let Some((left, right)) = concept.get_composition() {
-                        self.combine(&self.to_ast(left), &self.to_ast(right))
+                        self.combine(&self.to_ast(&left), &self.to_ast(&right))
                     } else {
                         self.snap_shot
-                            .new_syntax_from_concept_that_has_no_label_or_composition(&concept)
+                        .new_syntax_from_concept_that_has_no_label_or_composition(&concept)
                     }
                 }, |s| {
                     self.snap_shot
-                        .bind_concept_to_syntax(
-                            self.delta,
-                            Syntax::<C>::from(s),
-                            concept_id,
-                        )
+                    .bind_concept_to_syntax(
+                        self.delta.as_ref(),
+                        Syntax::<C>::from(s),
+                        concept_id,
+                    )
                 }).share();
                 self.caches.insert_syntax_tree(&concept, &syntax);
                 syntax
@@ -1100,13 +877,17 @@ where
         ast.get_concept()
             .and_also(&other.get_concept())
             .and_then(|(l, r)| {
-                let left_concept = self.snap_shot.read_concept(self.delta, *l);
+                let left_concept =
+                    self.snap_shot.read_concept(self.delta.as_ref(), *l);
                 left_concept
                     .find_as_hand_in_composition_with(*r, Hand::Left)
                     .map(|concept| {
                         let syntax = self.join(ast, other);
-                        self.snap_shot
-                            .bind_concept_to_syntax(self.delta, syntax, concept)
+                        self.snap_shot.bind_concept_to_syntax(
+                            self.delta.as_ref(),
+                            syntax,
+                            concept,
+                        )
                     })
             })
             .unwrap_or_else(|| self.join(ast, other))
@@ -1149,7 +930,7 @@ where
             Associativity::Right,
             |associativity_concept_id| {
                 let assoc_of_ast = self
-                    .combine(&self.to_ast(associativity_concept_id), ast)
+                    .combine(&self.to_ast(&associativity_concept_id), ast)
                     .share();
                 let maybe_concrete_concept_type = self
                     .reduce(&assoc_of_ast)
@@ -1177,18 +958,14 @@ where
                 }
             },
             |con| {
-                if let Some((left, right)) = self
-                    .snap_shot
-                    .read_concept(self.delta, con)
-                    .get_composition()
-                {
+                if let Some((left, right)) = self.composition_of_concept(&con) {
                     self.combine(
-                        &self.expand(&self.to_ast(left)),
-                        &self.expand(&self.to_ast(right)),
+                        &self.expand(&self.to_ast(&left)),
+                        &self.expand(&self.to_ast(&right)),
                     )
                     .share()
                 } else {
-                    self.to_ast(con)
+                    self.to_ast(&con)
                 }
             },
         )
@@ -1206,7 +983,7 @@ where
 
         self.concrete_concept_id(ConcreteConceptType::GreaterThan).map_or((Comparison::Incomparable, ComparisonReason::NoGreaterThanConcept), |greater_than_concept_id|
         {
-            let greater_than_syntax = self.to_ast(greater_than_concept_id);
+            let greater_than_syntax = self.to_ast(&greater_than_concept_id);
             let comparing_syntax = self.combine(
                 some_syntax,
                 &self.combine(&greater_than_syntax, another_syntax).share(),
@@ -1217,60 +994,62 @@ where
             );
             let mut reason = None;
             let mut reversed_reason = None;
+            let determine_concrete_type_whilst_noting_reversed_reason =|(reversed_comparison, local_reversed_reason)| {
+                reversed_reason = local_reversed_reason;
+                self.concrete_type_of_ast(&reversed_comparison)
+            };
+            let determine_concrete_type_whilst_noting_reason =|(syntax_comparison, local_reason)| {
+                reason = local_reason;
+                self.concrete_type_of_ast(&syntax_comparison)
+            };
             let cache = <C as ContextCache>::SharedReductionCache::default();
             (
                 match (
                     if self.syntax_evaluating.contains(&comparing_syntax) {
                         self.caches
-                            .get_reduction_or_else(
-                                &comparing_syntax.share(),
-                                || None,
-                            )
-                            .map(|(s, r)| (s, Some(r)))
+                        .get_reduction_or_else(
+                            &comparing_syntax.share(),
+                            || None,
+                        )
+                        .map(|(s, r)| (s, Some(r)))
                     } else {
-                        let mut context_search = self.spawn(&cache);
+                        let mut context_search = self.spawn(&cache, self.delta.clone());
                         let comparing_syntax = comparing_syntax.share();
                         context_search
-                            .syntax_evaluating
-                            .insert(comparing_syntax.clone());
-                        Some(
+                        .syntax_evaluating
+                        .insert(comparing_syntax.clone());
+                    Some(
                             context_search
-                                .recursively_reduce(&comparing_syntax),
+                            .recursively_reduce(&comparing_syntax),
                         )
                     }
                     .and_then(
-                        |(syntax_comparison, local_reason)| {
-                            reason = local_reason;
-                            self.concrete_type_of_ast(&syntax_comparison)
-                        },
+                        determine_concrete_type_whilst_noting_reason,
                     ),
                     if self
                         .syntax_evaluating
                         .contains(&comparing_reversed_syntax)
-                    {
-                        self.caches
+                        {
+                            self.caches
                             .get_reduction_or_else(
                                 &comparing_reversed_syntax.share(),
                                 || None,
                             )
                             .map(|(s, r)| (s, Some(r)))
-                    } else {
-                        let mut context_search = self.spawn(&cache);
+                        } else {
+                            let mut context_search = self.spawn(&cache, self.delta.clone());
                         let comparing_reversed_syntax =
-                            comparing_reversed_syntax.share();
+                        comparing_reversed_syntax.share();
                         context_search
-                            .syntax_evaluating
-                            .insert(comparing_reversed_syntax.clone());
-                        Some(
-                            context_search
-                                .recursively_reduce(&comparing_reversed_syntax),
-                        )
-                    }
-                    .and_then(
-                        |(reversed_comparison, local_reversed_reason)| {
-                            reversed_reason = local_reversed_reason;
-                            self.concrete_type_of_ast(&reversed_comparison)
-                        },
+                        .syntax_evaluating
+                        .insert(comparing_reversed_syntax.clone());
+                    Some(
+                        context_search
+                        .recursively_reduce(&comparing_reversed_syntax),
+                    )
+                }
+                .and_then(
+                        determine_concrete_type_whilst_noting_reversed_reason,
                     ),
                 ) {
                     (
@@ -1281,7 +1060,7 @@ where
                         Some(ConcreteConceptType::True),
                         Some(ConcreteConceptType::True),
                     ) => {
-                        panic!("{:#?} is both greater than and less than {:#?}!\nReason: {:#?}\n Reversed reason: {:#?}", some_syntax, another_syntax, reason, reversed_reason);
+                        panic!("{some_syntax:#?} is both greater than and less than {another_syntax:#?}!\nReason: {reason:#?}\n Reversed reason: {reversed_reason:#?}");
                     },
                     (Some(ConcreteConceptType::True), _) => {
                         Comparison::GreaterThan
@@ -1302,21 +1081,165 @@ where
         })
     }
 
-    pub fn spawn(
-        &self,
-        cache: &<C as ContextCache>::SharedReductionCache,
-    ) -> Self {
-        Self {
+    pub fn spawn<'b, 'c>(
+        &'b self,
+        cache: &'c <C as ContextCache>::SharedReductionCache,
+        delta: D,
+    ) -> Self
+    where
+        VML: VariableMaskList,
+    {
+        ContextSearch::<'s, 'v> {
+            concept_inferring: self.concept_inferring.clone(),
             bound_variable_syntax: self.bound_variable_syntax,
             caches: self.caches.spawn(cache),
-            delta: self.delta,
+            delta,
             snap_shot: self.snap_shot,
             syntax_evaluating: self.syntax_evaluating.clone(),
             variable_mask: self.variable_mask.clone(),
+            phantom: self.phantom,
+            phantom2: self.phantom2,
         }
     }
 
+    /// Checks if the concepts in the composition of `generalisation` matches the corresponding nodes in `ast` or is a free variable
+    /// returns `None` if does not match otherwise returns a `HashMap` mapping free variable concept IDs to the matching nodes in `ast`
+    /// TODO: refactor into method on struct with `self.snap_shot`, self.delta, `self.bound_variable_syntax` and `self.variable_mask`
+    pub fn check_generalisation(
+        &self,
+        ast: &SharedSyntax<C>,
+        generalisation: &S::ConceptId,
+    ) -> Option<VariableMask<Syntax<C>>> {
+        (self.is_free_variable(generalisation)
+            && !self.bound_variable_syntax.contains(ast)
+            && !ast.get_concept().map_or(false, |c| {
+                self.snap_shot
+                    .read_concept(self.delta.as_ref(), c)
+                    .bounded_variable()
+            }))
+        .then(|| {
+            if let Some((gl, gr)) = self.composition_of_concept(generalisation)
+            {
+                let (l, r) = ast.get_expansion()?;
+                match (self.is_free_variable(&gl), self.is_free_variable(&gr)) {
+                    (true, true) => self
+                        .check_generalisation(&l, &gl)
+                        .and_also_move(self.check_generalisation(&r, &gr))
+                        .and_then(|(lm, mut rm)| {
+                            for (lmk, lmv) in lm {
+                                if let Some(rmv) = rm.get(&lmk) {
+                                    if rmv != &lmv {
+                                        return None;
+                                    }
+                                } else {
+                                    rm.insert(lmk, lmv);
+                                }
+                            }
+                            Some(rm)
+                        }),
+                    (true, false) if r.get_concept() == Some(gr) => {
+                        self.check_generalisation(&l, &gl)
+                    },
+                    (false, true) if l.get_concept() == Some(gl) => {
+                        self.check_generalisation(&r, &gr)
+                    },
+                    (false, false)
+                        if l.get_concept() == Some(gl)
+                            && r.get_concept() == Some(gr) =>
+                    {
+                        Some(hashmap! {})
+                    },
+                    _ => None,
+                }
+            } else {
+                Some(hashmap! {*generalisation => ast.clone()})
+            }
+        })
+        .flatten()
+    }
+
+    // TODO: move to separate struct that just has self.delta, self.snap_shot and self.variable_mask
+    fn is_leaf_variable(&self, lv: &S::ConceptId) -> bool {
+        self.is_free_variable(lv) && self.is_leaf_concept(lv)
+    }
+
+    // TODO: move to separate struct that just has self.delta, self.snap_shot and self.variable_mask
+    fn is_free_variable(&self, v: &S::ConceptId) -> bool {
+        self.snap_shot.read_concept(self.delta.as_ref(), *v).free_variable()
+            && self
+                .variable_mask
+                .tail()
+                .map_or(true, |vml| vml.get(*v).is_none()) // A hack required to prevent stack overflow
+    }
+
+    // TODO: move to separate struct that just has self.delta and self.snap_shot
+    fn is_leaf_concept(&self, l: &S::ConceptId) -> bool {
+        self.composition_of_concept(l).is_none()
+    }
+
+    // TODO: move to separate struct that just has self.delta and self.snap_shot
+    fn equivalent_concepts_to(
+        &self,
+        equivalent_id: &S::ConceptId,
+    ) -> HashSet<S::ConceptId> {
+        let equivalent_concept =
+            self.snap_shot.read_concept(self.delta.as_ref(), *equivalent_id);
+        // TODO handle case when a concept implicitly reduces to `equivalent_concept`
+        let mut equivalence_set: HashSet<S::ConceptId> =
+            equivalent_concept.find_what_reduces_to_it().collect();
+        equivalence_set.insert(*equivalent_id);
+        equivalence_set
+    }
+
+    // TODO: move to separate struct that just has self.delta and self.snap_shot
+    fn composition_of_concept(
+        &self,
+        composition_id: &S::ConceptId,
+    ) -> Option<(S::ConceptId, S::ConceptId)> {
+        self.snap_shot
+            .read_concept(self.delta.as_ref(), *composition_id)
+            .get_composition()
+    }
+
+    // TODO: refactor into method of struct with access to self.snap_shot and self.delta
+    fn is_concrete_type(
+        &self,
+        cct: ConcreteConceptType,
+        concept_id: &S::ConceptId,
+    ) -> Option<S::ConceptId> {
+        if Some(cct) == self.concrete_type(concept_id) {
+            Some(*concept_id)
+        } else {
+            None
+        }
+    }
+
+    // TODO: refactor into method of struct with access to self.snap_shot and self.delta
+    fn concrete_type(
+        &self,
+        concept_id: &S::ConceptId,
+    ) -> Option<ConcreteConceptType> {
+        self.snap_shot.concrete_concept_type(self.delta.as_ref(), *concept_id)
+    }
+
+    // TODO: refactor into method of struct with access to self.snap_shot and self.delta
+    fn concrete_type_of_ast(
+        &self,
+        ast: &SharedSyntax<C>,
+    ) -> Option<ConcreteConceptType> {
+        ast.get_concept().and_then(|c| self.concrete_type(&c))
+    }
+
+    // TODO: refactor into method of struct with access to self.snap_shot and self.delta
+    fn concrete_concept_id(
+        &self,
+        cct: ConcreteConceptType,
+    ) -> Option<S::ConceptId> {
+        self.snap_shot.concrete_concept_id(self.delta.as_ref(), cct)
+    }
+
     /// Error if `variable_mask` is already included
+    /// TODO: refactor as method on `VariableMaskList`
     fn insert_variable_mask(
         &mut self,
         variable_mask: VariableMask<Syntax<C>>,
@@ -1326,6 +1249,7 @@ where
         Ok(())
     }
 
+    // TODO: move to separate struct with access to self.snap_shot, self.bound_variable_syntax, self.delta and self.caches
     fn contains_bound_variable_syntax(&self, syntax: &SharedSyntax<C>) -> bool {
         self.caches.remember_if_contains_bound_variable_syntax_or_else(
             syntax,
@@ -1340,7 +1264,7 @@ where
                 self.bound_variable_syntax.contains(syntax)
                     || syntax.get_concept().map_or(false, |c| {
                         self.snap_shot
-                            .read_concept(self.delta, c)
+                            .read_concept(self.delta.as_ref(), c)
                             .bounded_variable()
                     })
             },
@@ -1354,8 +1278,7 @@ pub struct Example<Syntax: SyntaxTree, RR: ReductionReason> {
     substitutions: HashMap<Syntax::SharedSyntax, Match<Syntax, RR>>,
 }
 
-type MaybeReducedSyntaxWithReason<RR> =
-    (ReductionReasonSharedSyntax<RR>, Option<RR>);
+type MaybeReducedSyntaxWithReason<RR> = (RRSharedSyntax<RR>, Option<RR>);
 
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub struct Match<Syntax: SyntaxTree, RR: ReductionReason> {
@@ -1384,15 +1307,21 @@ pub enum ComparisonReason<RR: ReductionReason> {
     NoGreaterThanConcept,
 }
 
-impl<'a, S, C, SDCD, VML> From<ContextReferences<'a, S, C, SDCD>>
-    for ContextSearch<'a, S, C, SDCD, VML>
+impl<'c, 's, 'v, S, C, SDCD, VML, D, CCI: MixedConcept>
+    From<ContextReferences<'c, 's, 'v, S, C, D>>
+    for ContextSearch<'s, 'v, S, C, VML, SDCD, D, CCI>
 where
+    S: SnapShotReader<SDCD, ConceptId = CCI> + Sync + std::fmt::Debug,
+    Self: Iteration<ConceptId = S::ConceptId, Syntax = Syntax<C>>,
     C: ContextCache,
-    S: SnapShotReader<SDCD>,
+    Syntax<C>: SyntaxTree<ConceptId = S::ConceptId>,
     SDCD: Clone
         + AsRef<DirectConceptDelta<S::ConceptId>>
-        + From<DirectConceptDelta<S::ConceptId>>,
-    VML: VariableMaskList,
+        + From<DirectConceptDelta<S::ConceptId>>
+        + Debug,
+    VML: VariableMaskList<Syntax = Syntax<C>>,
+    D: AsRef<NestedDelta<S::ConceptId, SDCD, D>>
+        + SharedDelta<NestedDelta = NestedDelta<S::ConceptId, SDCD, D>>,
 {
     fn from(
         ContextReferences {
@@ -1400,42 +1329,27 @@ where
             delta,
             cache,
             bound_variable_syntax,
-        }: ContextReferences<'a, S, C, SDCD>,
-    ) -> ContextSearch<'a, S, C, SDCD, VML> {
-        // simple_logger::init().unwrap_or(());
-        ContextSearch::<'a> {
+        }: ContextReferences<'c, 's, 'v, S, C, D>,
+    ) -> Self {
+        Self {
+            concept_inferring: HashSet::default(),
             bound_variable_syntax,
             snap_shot,
             variable_mask: VML::from(hashmap! {}).into(),
             delta,
             caches: cache.clone(),
             syntax_evaluating: hashset! {},
+            phantom: PhantomData,
+            phantom2: PhantomData,
         }
     }
 }
 
-pub struct ContextReferences<'a, S, C, SDCD>
-where
-    S: SnapShotReader<SDCD>,
-    C: ContextCache,
-    SDCD: Clone
-        + AsRef<DirectConceptDelta<S::ConceptId>>
-        + From<DirectConceptDelta<S::ConceptId>>,
-{
-    pub snap_shot: &'a S,
-    pub delta: &'a ContextDelta<S::ConceptId, SDCD>,
-    pub cache: &'a C,
-    pub bound_variable_syntax: &'a HashSet<SharedSyntax<C>>,
-}
-
-pub trait Iteration {
-    type Syntax: SyntaxTree;
-    type ConceptId;
-    fn filter_generalisations_from_candidates(
-        &self,
-        example: &<Self::Syntax as SyntaxTree>::SharedSyntax,
-        candidates: HashSet<Self::ConceptId>,
-    ) -> Generalisations<Self::Syntax>;
+pub struct ContextReferences<'c, 's, 'v, S, C: ContextCache, D> {
+    pub snap_shot: &'s S,
+    pub delta: D,
+    pub cache: &'c C,
+    pub bound_variable_syntax: &'v HashSet<SharedSyntax<C>>,
 }
 
 pub type Generalisations<Syntax> =
