@@ -2,11 +2,13 @@ use super::Syntax;
 use crate::{
     ast::SyntaxTree,
     concepts::{Concept, ConcreteConceptType, SpecificPart},
-    context_delta::ContextDelta,
+    context_delta::NestedDelta,
     context_search::{ContextReferences, ContextSearch},
     context_search_test::ReductionReason,
     mock_snap_shot::{ConceptId, MockSnapShot},
-    multi_threaded::{MultiThreadedContextCache, SharedDirectConceptDelta},
+    multi_threaded::{
+        MultiThreadedContextCache, SharedContextDelta, SharedDirectConceptDelta,
+    },
 };
 use maplit::{hashmap, hashset};
 use std::collections::HashMap;
@@ -14,19 +16,21 @@ use std::collections::HashMap;
 #[test]
 fn basic_existence() {
     let snapshot = MockSnapShot::new_test_case(&concepts(), &labels());
-    let delta =
-        ContextDelta::<ConceptId, SharedDirectConceptDelta<ConceptId>>::default(
-        );
+    let delta = NestedDelta::<
+        ConceptId,
+        SharedDirectConceptDelta<ConceptId>,
+        SharedContextDelta<ConceptId>,
+    >::default();
     let cache = MultiThreadedContextCache::default();
     let variable_syntax = Syntax::from("_x_").share();
     let bound_variables = hashset! {variable_syntax.clone()};
     let context_search = ContextSearch::from(ContextReferences {
         snap_shot: &snapshot,
-        delta: &delta,
+        delta: SharedContextDelta(delta.into()),
         cache: &cache,
         bound_variable_syntax: &bound_variables,
     });
-    let exists_such_that_syntax = context_search.to_ast(0);
+    let exists_such_that_syntax = context_search.to_ast(&0);
     let variable_exists_such_that_variable_is_true_syntax = context_search
         .combine(
             &context_search
@@ -35,21 +39,21 @@ fn basic_existence() {
             &variable_syntax,
         )
         .into();
-    let true_syntax = context_search.to_ast(1);
+    let true_syntax = context_search.to_ast(&1);
     assert_eq!(
         context_search
             .reduce(&variable_exists_such_that_variable_is_true_syntax),
         Some((
-            true_syntax.clone(),
+            true_syntax,
             ReductionReason::Existence {
                 generalisation: variable_syntax.clone(),
-                substitutions: hashmap! {variable_syntax => context_search.to_ast(2)},
+                substitutions: hashmap! {variable_syntax => context_search.to_ast(&2)},
             }
         ))
     );
 }
 
-fn labels() -> HashMap<ConceptId, &'static str> {
+fn labels() -> HashMap<usize, &'static str> {
     hashmap! {
         0 => "exists_such_that",
         1 => "true",
@@ -57,11 +61,11 @@ fn labels() -> HashMap<ConceptId, &'static str> {
     }
 }
 
-fn concepts() -> [Concept<ConceptId>; 3] {
+fn concepts() -> [Concept<usize>; 3] {
     let exists_such_that_concept =
         (ConcreteConceptType::ExistsSuchThat, 0).into();
     let mut true_concept = (ConcreteConceptType::True, 1).into();
-    let mut abstract_concept: Concept<ConceptId> =
+    let mut abstract_concept: Concept<usize> =
         (SpecificPart::default(), 2).into();
     abstract_concept.make_reduce_to(&mut true_concept);
     [exists_such_that_concept, true_concept, abstract_concept]
