@@ -1,12 +1,14 @@
 use crate::{
     concepts::{Concept, ConcreteConceptType, SpecificPart},
-    context_delta::ContextDelta,
+    context_delta::NestedDelta,
     context_search::{
         Comparison, ComparisonReason, ContextReferences, ContextSearch,
     },
     context_search_test::ReductionReason,
-    mock_snap_shot::{ConceptId, MockSnapShot},
-    multi_threaded::{MultiThreadedContextCache, SharedDirectConceptDelta},
+    mock_snap_shot::MockSnapShot,
+    multi_threaded::{
+        MultiThreadedContextCache, SharedContextDelta, SharedDirectConceptDelta,
+    },
 };
 use maplit::{hashmap, hashset};
 use std::collections::HashMap;
@@ -15,31 +17,31 @@ use std::collections::HashMap;
 fn comparison_existence_implication_rule_test() {
     let context_cache = MultiThreadedContextCache::default();
     let context_delta =
-        ContextDelta::<_, SharedDirectConceptDelta<_>>::default();
+        NestedDelta::<_, SharedDirectConceptDelta<_>, _>::default();
     let context_snap_shot = MockSnapShot::new_test_case(&concepts(), &labels());
     let bound_variables = hashset! {};
     let context_search = ContextSearch::from(ContextReferences {
         snap_shot: &context_snap_shot,
-        delta: &context_delta,
+        delta: SharedContextDelta(context_delta.into()),
         cache: &context_cache,
         bound_variable_syntax: &bound_variables,
     });
-    let a_syntax = context_search.to_ast(21);
-    let c_syntax = context_search.to_ast(23);
+    let a_syntax = context_search.to_ast(&21);
+    let c_syntax = context_search.to_ast(&23);
     let variable_mask = hashmap! {
-        4 => a_syntax.clone(), // x=a
-        6 => c_syntax.clone() //z=c
+        4.into() => a_syntax.clone(), // x=a
+        6.into() => c_syntax.clone() //z=c
     };
     let comparison_reason = Some(ReductionReason::Rule{
-        generalisation: context_search.to_ast(18),
+        generalisation: context_search.to_ast(&18),
         variable_mask: variable_mask.clone(),
         reason: ReductionReason::Inference{
             implication: context_search.substitute(
-                &context_search.to_ast(20), &variable_mask
+                &context_search.to_ast(&20), &variable_mask
             ),
             reason: ReductionReason::Existence{
-                substitutions: hashmap!{context_search.to_ast(5) => context_search.to_ast(22)},
-                generalisation: context_search.substitute(&context_search.to_ast(15), &variable_mask),
+                substitutions: hashmap!{context_search.to_ast(&5) => context_search.to_ast(&22)},
+                generalisation: context_search.substitute(&context_search.to_ast(&15), &variable_mask),
             }.into()
         }.into()
     });
@@ -63,14 +65,14 @@ fn comparison_existence_implication_rule_test() {
             }
         )
     );
-    let d_syntax = context_search.to_ast(28);
+    let d_syntax = context_search.to_ast(&28);
     assert_eq!(
         context_search.compare(&a_syntax, &d_syntax).0,
         Comparison::GreaterThan
-    )
+    );
 }
 
-fn labels() -> HashMap<ConceptId, &'static str> {
+fn labels() -> HashMap<usize, &'static str> {
     hashmap! {
         0 => "true",
         1 => ">",
@@ -87,19 +89,21 @@ fn labels() -> HashMap<ConceptId, &'static str> {
     }
 }
 
-fn concepts() -> [Concept<ConceptId>; 31] {
+#[allow(clippy::too_many_lines)]
+fn concepts() -> [Concept<usize>; 31] {
     let mut true_concept = (ConcreteConceptType::True, 0).into();
     let mut greater_than_concept = (ConcreteConceptType::GreaterThan, 1).into();
     let mut exists_such_that_concept =
         (ConcreteConceptType::ExistsSuchThat, 2).into();
     let mut implication_concept = (ConcreteConceptType::Implication, 3).into();
-    let mut x = (SpecificPart::free_variable(), 4).into();
-    let mut y = (SpecificPart::bound_variable(), 5).into();
-    let mut z = (SpecificPart::free_variable(), 6).into();
+    let mut concept_x: Concept<usize> =
+        (SpecificPart::free_variable(), 4).into();
+    let mut concept_y = (SpecificPart::bound_variable(), 5).into();
+    let mut concept_z = (SpecificPart::free_variable(), 6).into();
     let mut greater_than_z =
-        Concept::composition_of(7, &mut greater_than_concept, &mut z);
+        Concept::composition_of(7, &mut greater_than_concept, &mut concept_z);
     let mut y_greater_than_z =
-        Concept::composition_of(8, &mut y, &mut greater_than_z);
+        Concept::composition_of(8, &mut concept_y, &mut greater_than_z);
     let mut and_concept = (SpecificPart::default(), 9).into();
     let mut and_true =
         Concept::composition_of(10, &mut and_concept, &mut true_concept);
@@ -109,16 +113,19 @@ fn concepts() -> [Concept<ConceptId>; 31] {
     let mut and_y_greater_than_z =
         Concept::composition_of(12, &mut and_concept, &mut y_greater_than_z);
     let mut greater_than_y =
-        Concept::composition_of(13, &mut greater_than_concept, &mut y);
+        Concept::composition_of(13, &mut greater_than_concept, &mut concept_y);
     let mut x_greater_than_y =
-        Concept::composition_of(14, &mut x, &mut greater_than_y);
+        Concept::composition_of(14, &mut concept_x, &mut greater_than_y);
     let mut x_greater_than_y_and_y_greater_than_z = Concept::composition_of(
         15,
         &mut x_greater_than_y,
         &mut and_y_greater_than_z,
     );
-    let mut y_exists_such_that =
-        Concept::composition_of(16, &mut y, &mut exists_such_that_concept);
+    let mut y_exists_such_that = Concept::composition_of(
+        16,
+        &mut concept_y,
+        &mut exists_such_that_concept,
+    );
     let mut y_exists_such_that_x_greater_than_y_and_y_greater_than_z =
         Concept::composition_of(
             17,
@@ -126,7 +133,7 @@ fn concepts() -> [Concept<ConceptId>; 31] {
             &mut x_greater_than_y_and_y_greater_than_z,
         );
     let mut x_greater_than_z =
-        Concept::composition_of(18, &mut x, &mut greater_than_z);
+        Concept::composition_of(18, &mut concept_x, &mut greater_than_z);
     let mut implies_x_greater_than_z = Concept::composition_of(
         19,
         &mut implication_concept,
@@ -140,33 +147,33 @@ fn concepts() -> [Concept<ConceptId>; 31] {
             &mut implies_x_greater_than_z,
         );
     y_exists_such_that_x_greater_than_y_and_y_greater_than_z_implies_x_greater_than_z.make_reduce_to(&mut true_concept);
-    let mut a = (SpecificPart::default(), 21).into();
-    let mut b = (SpecificPart::default(), 22).into();
-    let mut c = (SpecificPart::default(), 23).into();
+    let mut concept_a = (SpecificPart::default(), 21).into();
+    let mut concept_b = (SpecificPart::default(), 22).into();
+    let mut concept_c = (SpecificPart::default(), 23).into();
     let mut greater_than_b =
-        Concept::composition_of(24, &mut greater_than_concept, &mut b);
+        Concept::composition_of(24, &mut greater_than_concept, &mut concept_b);
     let mut a_greater_than_b =
-        Concept::composition_of(25, &mut a, &mut greater_than_b);
+        Concept::composition_of(25, &mut concept_a, &mut greater_than_b);
     a_greater_than_b.make_reduce_to(&mut true_concept);
     let mut greater_than_c =
-        Concept::composition_of(26, &mut greater_than_concept, &mut c);
+        Concept::composition_of(26, &mut greater_than_concept, &mut concept_c);
     let mut b_greater_than_c =
-        Concept::composition_of(27, &mut b, &mut greater_than_c);
+        Concept::composition_of(27, &mut concept_b, &mut greater_than_c);
     b_greater_than_c.make_reduce_to(&mut true_concept);
-    let mut d = (SpecificPart::default(), 28).into();
+    let mut concept_d: Concept<usize> = (SpecificPart::default(), 28).into();
     let mut greater_than_d =
-        Concept::composition_of(29, &mut greater_than_concept, &mut d);
+        Concept::composition_of(29, &mut greater_than_concept, &mut concept_d);
     let mut c_greater_than_d =
-        Concept::composition_of(30, &mut c, &mut greater_than_d);
+        Concept::composition_of(30, &mut concept_c, &mut greater_than_d);
     c_greater_than_d.make_reduce_to(&mut true_concept);
     [
         true_concept,
         greater_than_concept,
         exists_such_that_concept,
         implication_concept,
-        x,
-        y,
-        z,
+        concept_x,
+        concept_y,
+        concept_z,
         greater_than_z,
         y_greater_than_z,
         and_concept,
@@ -181,14 +188,14 @@ fn concepts() -> [Concept<ConceptId>; 31] {
         x_greater_than_z,
         implies_x_greater_than_z,
         y_exists_such_that_x_greater_than_y_and_y_greater_than_z_implies_x_greater_than_z,
-        a,
-        b,
-        c,
+        concept_a,
+        concept_b,
+        concept_c,
         greater_than_b,
         a_greater_than_b,
         greater_than_c,
         b_greater_than_c,
-        d,
+        concept_d,
         greater_than_d,
         c_greater_than_d
     ]
