@@ -2,18 +2,18 @@ use super::Syntax;
 use crate::{
     ast::SyntaxTree,
     concepts::{Concept, ConcreteConceptType, SpecificPart},
-    context_delta::ContextDelta,
+    context_delta::NestedDelta,
     context_search::{ContextReferences, ContextSearch},
-    mock_snap_shot::{ConceptId, MockSnapShot},
+    mock_snap_shot::MockSnapShot,
     multi_threaded::{
         MultiThreadedContextCache, MultiThreadedReductionReason,
-        SharedDirectConceptDelta,
+        SharedContextDelta, SharedDirectConceptDelta,
     },
 };
 use maplit::{hashmap, hashset};
 use std::collections::HashMap;
 
-fn labels() -> HashMap<ConceptId, &'static str> {
+fn labels() -> HashMap<usize, &'static str> {
     hashmap! {
         0 => "=>",
         1 => "true",
@@ -32,34 +32,34 @@ type ReductionReason = MultiThreadedReductionReason<Syntax>;
 fn existence_inference_rule() {
     let context_cache = MultiThreadedContextCache::default();
     let context_delta =
-        ContextDelta::<_, SharedDirectConceptDelta<_>>::default();
+        NestedDelta::<_, SharedDirectConceptDelta<_>, _>::default();
     let context_snap_shot = MockSnapShot::new_test_case(&concepts(), &labels());
     let bound_variables = hashset! {};
     let context_search = ContextSearch::from(ContextReferences {
         snap_shot: &context_snap_shot,
-        delta: &context_delta,
+        delta: SharedContextDelta(context_delta.into()),
         cache: &context_cache,
         bound_variable_syntax: &bound_variables,
     });
     let example_syntax =
-        Syntax::new_pair(context_search.to_ast(7), context_search.to_ast(10))
+        Syntax::new_pair(context_search.to_ast(&7), context_search.to_ast(&10))
             .share();
     let true_syntax = Syntax::from("true").bind_nonquantifier_concept(1);
-    let variable_mask = hashmap! {9 => context_search.to_ast(7)};
+    let variable_mask = hashmap! {9.into() => context_search.to_ast(&7)};
     assert_eq!(
         context_search.reduce(&example_syntax),
         Some((
             true_syntax.into(),
             ReductionReason::Rule {
-                generalisation: context_search.to_ast(2),
+                generalisation: context_search.to_ast(&2),
                 variable_mask: variable_mask.clone(),
                 reason: ReductionReason::Inference {
                     implication: context_search
-                        .substitute(&context_search.to_ast(6), &variable_mask),
+                        .substitute(&context_search.to_ast(&6), &variable_mask),
                     reason: ReductionReason::Existence{
                         generalisation: context_search
-                            .substitute(&context_search.to_ast(13), &variable_mask),
-                        substitutions: hashmap!{context_search.to_ast(11) => context_search.to_ast(3)},
+                            .substitute(&context_search.to_ast(&13), &variable_mask),
+                        substitutions: hashmap!{context_search.to_ast(&11) => context_search.to_ast(&3)},
                     }
                     .into()
                 }
@@ -69,7 +69,7 @@ fn existence_inference_rule() {
     );
 }
 
-fn concepts() -> [Concept<ConceptId>; 17] {
+fn concepts() -> [Concept<usize>; 17] {
     let mut implication_concept = (ConcreteConceptType::Implication, 0).into();
     let mut true_concept = (ConcreteConceptType::True, 1).into();
     let mut concept_a = (SpecificPart::default(), 3).into();
