@@ -6,8 +6,11 @@ mod basic_reduction;
 mod basic_rule;
 mod comparison_existence_implication_rule;
 mod existence_inference_rule;
+mod implied_reduction_via_implication_chain;
 mod inference_rule;
+mod inferred_negation;
 mod nested_composition_rule;
+mod not;
 
 use crate::{
     concepts::{Concept, ConceptTrait},
@@ -15,13 +18,13 @@ use crate::{
     multi_threaded::{MultiThreadedReductionReason, MultiThreadedSyntaxTree},
 };
 
-pub fn check_order(concepts: &[Concept<ConceptId>]) -> Vec<Concept<ConceptId>> {
+pub fn check_order(concepts: &[Concept<usize>]) -> Vec<Concept<ConceptId>> {
     concepts
         .iter()
         .enumerate()
         .map(|(i, c)| {
             assert_eq!(i, c.id());
-            c.clone()
+            c.clone().into()
         })
         .collect::<Vec<Concept<ConceptId>>>()
 }
