@@ -2,16 +2,18 @@ use super::Syntax;
 use crate::{
     ast::SyntaxTree,
     concepts::{Concept, ConcreteConceptType, SpecificPart},
-    context_delta::ContextDelta,
+    context_delta::NestedDelta,
     context_search::{ContextReferences, ContextSearch},
     context_search_test::ReductionReason,
     mock_snap_shot::{ConceptId, MockSnapShot},
-    multi_threaded::{MultiThreadedContextCache, SharedDirectConceptDelta},
+    multi_threaded::{
+        MultiThreadedContextCache, SharedContextDelta, SharedDirectConceptDelta,
+    },
 };
 use maplit::{hashmap, hashset};
 use std::collections::HashMap;
 
-fn concepts() -> [Concept<ConceptId>; 8] {
+fn concepts() -> [Concept<usize>; 8] {
     let mut concrete_concept = (ConcreteConceptType::True, 0).into();
     let mut left_concept = (SpecificPart::default(), 2).into();
     let mut right_left_concept = (SpecificPart::default(), 3).into();
@@ -42,7 +44,7 @@ fn concepts() -> [Concept<ConceptId>; 8] {
     ]
 }
 
-fn labels() -> HashMap<ConceptId, &'static str> {
+fn labels() -> HashMap<usize, &'static str> {
     hashmap! {
         0 => "concrete",
         2 => "left",
@@ -55,12 +57,12 @@ fn labels() -> HashMap<ConceptId, &'static str> {
 fn basic_rule() {
     let snapshot = MockSnapShot::new_test_case(&concepts(), &labels());
     let delta =
-        ContextDelta::<_, SharedDirectConceptDelta<ConceptId>>::default();
+        NestedDelta::<_, SharedDirectConceptDelta<ConceptId>, _>::default();
     let cache = MultiThreadedContextCache::default();
     let bound_variable_syntax = hashset! {};
     let context_search = ContextSearch::from(ContextReferences {
         snap_shot: &snapshot,
-        delta: &delta,
+        delta: SharedContextDelta(delta.into()),
         cache: &cache,
         bound_variable_syntax: &bound_variable_syntax,
     });
@@ -80,12 +82,12 @@ fn basic_rule() {
     )
     .into();
 
-    assert_eq!(context_search.to_ast(0), concrete_syntax().into());
-    assert_eq!(context_search.to_ast(2), left_syntax);
+    assert_eq!(context_search.to_ast(&0), concrete_syntax().into());
+    assert_eq!(context_search.to_ast(&2), left_syntax);
 
     let reduction_reason = ReductionReason::Rule {
-        generalisation: context_search.to_ast(1),
-        variable_mask: hashmap! {4 => Syntax::from("random").into()},
+        generalisation: context_search.to_ast(&1),
+        variable_mask: hashmap! {4.into() => Syntax::from("random").into()},
         reason: ReductionReason::Explicit.into(),
     };
 
