@@ -1,4 +1,3 @@
-use super::Syntax;
 use crate::{
     ast::SyntaxTree,
     concepts::{Concept, ConcreteConceptType, SpecificPart},
@@ -10,50 +9,53 @@ use crate::{
     },
 };
 use maplit::{hashmap, hashset};
-use pretty_assertions::assert_eq;
 use std::collections::HashMap;
 
-fn concepts() -> [Concept<usize>; 15] {
+fn concepts() -> [Concept<usize>; 17] {
     let mut implication_concept = (ConcreteConceptType::Implication, 0).into();
     let mut true_concept = (ConcreteConceptType::True, 1).into();
-    let mut concept_a = (SpecificPart::default(), 3).into();
-    let mut example_concept = (SpecificPart::default(), 7).into();
-    let mut variable_concept = (SpecificPart::free_variable(), 9).into();
-    let mut concept_b = (SpecificPart::default(), 10).into();
-    let mut cause_concept =
-        Concept::composition_of(4, &mut concept_a, &mut variable_concept);
+    let mut variable = (SpecificPart::free_variable(), 3).into();
+    let mut not = (SpecificPart::default(), 7).into();
+    let mut reduction_concept = (ConcreteConceptType::Reduction, 4).into();
+    let mut false_concept = (ConcreteConceptType::False, 12).into();
+    let mut reduces_to_false =
+        Concept::composition_of(13, &mut reduction_concept, &mut false_concept);
     let mut result_concept =
-        Concept::composition_of(2, &mut variable_concept, &mut concept_b);
+        Concept::composition_of(2, &mut variable, &mut reduces_to_false);
     let mut implies_result_concept = Concept::composition_of(
         5,
         &mut implication_concept,
         &mut result_concept,
     );
+    let mut not_variable = Concept::composition_of(16, &mut not, &mut variable);
     let mut cause_implies_result_concept = Concept::composition_of(
         6,
-        &mut cause_concept,
+        &mut not_variable,
         &mut implies_result_concept,
     );
-    cause_implies_result_concept.make_reduce_to(&mut true_concept);
-    let mut example_composition =
-        Concept::composition_of(8, &mut concept_a, &mut example_concept);
-    example_composition.make_reduce_to(&mut true_concept);
+    let mut concept_c = (SpecificPart::default(), 14).into();
+    let mut not_concept_c =
+        Concept::composition_of(15, &mut not, &mut concept_c);
+    not_concept_c.make_reduce_to(&mut true_concept); // not c -> true
+    cause_implies_result_concept.make_reduce_to(&mut true_concept); // not _x_ => _x_ -> false
     [
         implication_concept,
         true_concept,
         result_concept,
-        concept_a,
-        cause_concept,
+        variable,
+        reduction_concept,
         implies_result_concept,
         cause_implies_result_concept,
-        example_concept,
-        example_composition,
-        variable_concept,
-        concept_b,
-        (ConcreteConceptType::Precedence, 11).into(),
-        (ConcreteConceptType::Associativity, 12).into(),
-        (ConcreteConceptType::Left, 13).into(),
-        (ConcreteConceptType::Right, 14).into(),
+        not,
+        (ConcreteConceptType::Precedence, 8).into(),
+        (ConcreteConceptType::Associativity, 9).into(),
+        (ConcreteConceptType::Left, 10).into(),
+        (ConcreteConceptType::Right, 11).into(),
+        false_concept,
+        reduces_to_false,
+        concept_c,
+        not_concept_c,
+        not_variable,
     ]
 }
 
@@ -61,19 +63,16 @@ fn labels() -> HashMap<usize, &'static str> {
     hashmap! {
         0 => "=>",
         1 => "true",
-        3 => "a",
-        7 => "example",
-        9 => "_x_",
-        10 => "b",
-        11 => "prec",
-        12 => "assoc",
-        13 => "left",
-        14 => "right"
+        3 => "_x_",
+        7 => "not",
+        4 => "->",
+        12 => "false",
+        14 => "c"
     }
 }
 
 #[test]
-fn inference_rule() {
+fn not() {
     let context_cache = MultiThreadedContextCache::default();
     let context_delta =
         NestedDelta::<_, SharedDirectConceptDelta<ConceptId>, _>::default();
@@ -85,11 +84,8 @@ fn inference_rule() {
         cache: &context_cache,
         bound_variable_syntax: &bound_variable_syntax,
     });
-    let example_syntax =
-        Syntax::new_pair(context_search.to_ast(&7), context_search.to_ast(&10));
-    let true_syntax = context_search.to_ast(&1);
-    assert_eq!(
-        context_search.reduce(&example_syntax.into()).unwrap().0,
-        true_syntax,
-    );
+    let (reduction, _) = context_search
+        .find_examples_of_inferred_reduction(&context_search.to_ast(&14))
+        .expect("Examples should be found");
+    assert_eq!(reduction.get_concept().unwrap(), 12.into());
 }
