@@ -2,11 +2,11 @@ use crate::{
     ast::impl_syntax_tree,
     context::Context as GenericContext,
     context_cache::impl_cache,
-    context_delta::DirectConceptDelta,
-    context_search::{
-        ContextSearch, Generalisations, Iteration as ContextSearchIteration,
-    },
+    context_delta::{DirectConceptDelta, NestedDelta, SharedDelta},
+    context_search::{ContextSearch, Generalisations},
     context_snap_shot::{ConceptId as ContextConceptId, ContextSnapShot},
+    iteration::Iteration as ContextSearchIteration,
+    mixed_concept::MixedConcept,
     snap_shot::Reader as SnapShotReader,
     variable_mask_list::impl_variable_mask_list,
 };
@@ -26,32 +26,85 @@ pub type Context = GenericContext<
     >,
     SharedDirectConceptDelta<ContextConceptId>,
     MultiThreadedVariableMaskList<MultiThreadedSyntaxTree<ContextConceptId>>,
+    SharedContextDelta<ContextConceptId>,
+    ContextConceptId,
 >;
 
 // Saves having to construct a new `Context` each time.
 lazy_static! {
-    pub static ref NEW_CONTEXT: Context = Context::new();
+    pub static ref NEW_CONTEXT: Context = Context::new().unwrap();
 }
 
-pub type SharedDirectConceptDelta<ConceptId> =
-    Arc<DirectConceptDelta<ConceptId>>;
+type MultiThreadedContextDelta<CCI> =
+    NestedDelta<CCI, SharedDirectConceptDelta<CCI>, SharedContextDelta<CCI>>;
 
-impl<'a, S, SDCD> ContextSearchIteration
+#[derive(Debug)]
+pub struct SharedContextDelta<CCI: MixedConcept>(
+    pub Arc<MultiThreadedContextDelta<CCI>>,
+);
+
+impl<CCI: MixedConcept> Clone for SharedContextDelta<CCI> {
+    fn clone(&self) -> Self {
+        Self(self.0.clone())
+    }
+}
+
+impl<CCI: MixedConcept> Default for SharedContextDelta<CCI> {
+    fn default() -> Self {
+        Self(MultiThreadedContextDelta::default().into())
+    }
+}
+
+impl<CCI: MixedConcept> SharedDelta for SharedContextDelta<CCI> {
+    type NestedDelta = MultiThreadedContextDelta<CCI>;
+
+    fn get_mut(&mut self) -> Option<&mut Self::NestedDelta> {
+        Arc::get_mut(&mut self.0)
+    }
+
+    fn from_nested(nested: Self::NestedDelta) -> Self {
+        Self(nested.into())
+    }
+
+    fn into_nested(self) -> crate::errors::ZiaResult<Self::NestedDelta> {
+        Arc::try_unwrap(self.0)
+            .map_err(|_| crate::ZiaError::MultiplePointersToDelta)
+    }
+
+    fn strong_count(&self) -> usize {
+        Arc::strong_count(&self.0)
+    }
+}
+
+impl<CCI: MixedConcept> AsRef<MultiThreadedContextDelta<CCI>>
+    for SharedContextDelta<CCI>
+{
+    fn as_ref(&self) -> &MultiThreadedContextDelta<CCI> {
+        self.0.as_ref()
+    }
+}
+
+pub type SharedDirectConceptDelta<CCI> = Arc<DirectConceptDelta<CCI>>;
+
+impl<'s, 'v, S, CCI: MixedConcept + Send + Sync> ContextSearchIteration
     for ContextSearch<
-        'a,
+        's,
+        'v,
         S,
         MultiThreadedContextCache<
             MultiThreadedReductionReason<MultiThreadedSyntaxTree<S::ConceptId>>,
         >,
-        SDCD,
         MultiThreadedVariableMaskList<MultiThreadedSyntaxTree<S::ConceptId>>,
+        SharedDirectConceptDelta<S::ConceptId>,
+        SharedContextDelta<S::ConceptId>,
+        CCI,
     >
 where
-    S: SnapShotReader<SDCD> + Sync + Debug,
-    SDCD: Clone
-        + AsRef<DirectConceptDelta<S::ConceptId>>
-        + From<DirectConceptDelta<S::ConceptId>>
-        + Sync,
+    S: SnapShotReader<SharedDirectConceptDelta<CCI>, ConceptId = CCI>
+        + Sync
+        + Debug,
+    for<'a> &'a std::collections::HashSet<CCI>:
+        rayon::iter::IntoParallelIterator<Item = &'a CCI>,
 {
     type ConceptId = S::ConceptId;
     type Syntax = MultiThreadedSyntaxTree<S::ConceptId>;
@@ -64,7 +117,7 @@ where
         candidates
             .par_iter()
             .filter_map(|gc| {
-                self.check_generalisation(example, *gc).and_then(|vm| {
+                self.check_generalisation(example, gc).and_then(|vm| {
                     if vm.is_empty() {
                         None
                     } else {
