@@ -2,11 +2,11 @@ use crate::{
     ast::impl_syntax_tree,
     context::Context as GenericContext,
     context_cache::impl_cache,
-    context_delta::DirectConceptDelta,
-    context_search::{
-        ContextSearch, Generalisations, Iteration as ContextSearchIteration,
-    },
+    context_delta::{DirectConceptDelta, NestedDelta, SharedDelta},
+    context_search::{ContextSearch, Generalisations},
     context_snap_shot::{ConceptId as ContextConceptId, ContextSnapShot},
+    errors::ZiaResult,
+    iteration::Iteration as ContextSearchIteration,
     snap_shot::Reader as SnapShotReader,
     variable_mask_list::impl_variable_mask_list,
 };
@@ -26,27 +26,64 @@ pub type Context = GenericContext<
     >,
     SharedDirectConceptDelta,
     SingleThreadedVariableMaskList<SingleThreadedSyntaxTree<ContextConceptId>>,
+    SharedContextDelta,
+    ContextConceptId,
 >;
+
+type SingleThreadedContextDelta =
+    NestedDelta<ContextConceptId, SharedDirectConceptDelta, SharedContextDelta>;
+
+#[derive(Clone, Default, Debug)]
+pub struct SharedContextDelta(Rc<SingleThreadedContextDelta>);
+
+impl SharedDelta for SharedContextDelta {
+    type NestedDelta = SingleThreadedContextDelta;
+
+    fn get_mut(&mut self) -> Option<&mut Self::NestedDelta> {
+        Rc::get_mut(&mut self.0)
+    }
+
+    fn from_nested(nested: Self::NestedDelta) -> Self {
+        Self(nested.into())
+    }
+
+    fn into_nested(self) -> ZiaResult<Self::NestedDelta> {
+        Rc::try_unwrap(self.0)
+            .map_err(|_| crate::ZiaError::MultiplePointersToDelta)
+    }
+
+    fn strong_count(&self) -> usize {
+        Rc::strong_count(&self.0)
+    }
+}
+
+impl AsRef<SingleThreadedContextDelta> for SharedContextDelta {
+    fn as_ref(&self) -> &SingleThreadedContextDelta {
+        self.0.as_ref()
+    }
+}
 
 type SharedDirectConceptDelta = Rc<DirectConceptDelta<ContextConceptId>>;
 
-impl<'a, S, SDCD> ContextSearchIteration
+impl<'s, 'v, S> ContextSearchIteration
     for ContextSearch<
-        'a,
+        's,
+        'v,
         S,
         SingleThreadedContextCache<
             SingleThreadedReductionReason<
                 SingleThreadedSyntaxTree<S::ConceptId>,
             >,
         >,
-        SDCD,
         SingleThreadedVariableMaskList<SingleThreadedSyntaxTree<S::ConceptId>>,
+        SharedDirectConceptDelta,
+        SharedContextDelta,
+        ContextConceptId,
     >
 where
-    S: SnapShotReader<SDCD, ConceptId = ContextConceptId> + Sync + Debug,
-    SDCD: Clone
-        + AsRef<DirectConceptDelta<ContextConceptId>>
-        + From<DirectConceptDelta<ContextConceptId>>,
+    S: SnapShotReader<SharedDirectConceptDelta, ConceptId = ContextConceptId>
+        + Sync
+        + Debug,
 {
     type ConceptId = S::ConceptId;
     type Syntax = SingleThreadedSyntaxTree<S::ConceptId>;
@@ -59,7 +96,7 @@ where
         candidates
             .iter()
             .filter_map(|gc| {
-                self.check_generalisation(example, *gc).and_then(|vm| {
+                self.check_generalisation(example, gc).and_then(|vm| {
                     if vm.is_empty() {
                         None
                     } else {
